+++ paper/paper.bib
@@ -4,7 +4,8 @@
       year={2015},
       eprint={1411.1607},
       archivePrefix={arXiv},
-      primaryClass={cs.MS}
+      primaryClass={cs.MS},
+      doi = {10.1137/141000671}
 }
 
 @article{juliafast,
@@ -135,7 +136,8 @@ archivePrefix = {arXiv},
        volume = {117},
         pages = {137-147},
        adsurl = {https://ui.adsabs.harvard.edu/abs/1996A&AS..117..137H},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+      doi = {10.1051/aas:1996146}
 }
 
 
