+++ paper/paper.bib
@@ -126,18 +126,16 @@ archivePrefix = {arXiv},
 
 
 
-@ARTICLE{Hamaker,
-       author = {{Hamaker}, J.~P. and {Bregman}, J.~D. and {Sault}, R.~J.},
-        title = "{Understanding radio polarimetry. I. Mathematical foundations.}",
-      journal = {\aaps},
-     keywords = {METHODS: ANALYTICAL, METHODS: DATA ANALYSIS, TECHNIQUES: INTERFEROMETERS, TECHNIQUES: POLARIMETERS},
-         year = 1996,
-        month = may,
-       volume = {117},
-        pages = {137-147},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/1996A&AS..117..137H},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System},
-      doi = {10.1051/aas:1996146}
+@article{Hamaker,
+	author = {{Hamaker, J. P.} and {Bregman, J. D.} and {Sault, R. J.}},
+	title = { Understanding radio polarimetry. I. Mathematical foundations },
+	DOI= "10.1051/aas:1996146",
+	url= "https://doi.org/10.1051/aas:1996146",
+	journal = {Astron. Astrophys. Suppl. Ser.},
+	year = 1996,
+	volume = 117,
+	number = 1,
+	pages = "137-147",
 }
 
 
