### 2022-03-23 #4262 @ptiede
I am not sure why the check references command is struggling with the 10.1051/aas:1996146 doi. It looks like it is ignoring the semi-colon when doing the check. The .bib file is correct.
