+++ src/observations/ehtim.jl
@@ -186,26 +186,240 @@ function extract_vis(obs)
     )
 end
 
+
+function closure_designmat(scancp::Scan{A,B,C}, scanvis) where {A,B,C<:StructArray{<:EHTClosurePhaseDatum}}
+    ant1, ant2, ant3 = baselines(scancp)
+    antvis1, antvis2 = baselines(scanvis)
+    design_mat = zeros(length(scancp), length(scanvis))
+    # fill in each row of the design matrix
+    for i in axes(design_mat, 2), j in axes(design_mat,1)
+        a1 = ant1[j]
+        a2 = ant2[j]
+        a3 = ant3[j]
+
+        # check leg 1
+        ((antvis1[i] == a1) & (antvis2[i] == a2)) && (design_mat[j,i] = 1.0)
+        ((antvis1[i] == a2) & (antvis2[i] == a1)) && (design_mat[j,i] = -1.0)
+
+        #check leg 2
+        ((antvis1[i] == a2) & (antvis2[i] == a3)) && (design_mat[j,i] = 1.0)
+        ((antvis1[i] == a3) & (antvis2[i] == a2)) && (design_mat[j,i] = -1.0)
+
+        #check leg 3
+        ((antvis1[i] == a3) & (antvis2[i] == a1)) && (design_mat[j,i] = 1.0)
+        ((antvis1[i] == a1) & (antvis2[i] == a3)) && (design_mat[j,i] = -1.0)
+    end
+    return design_mat
+end
+
+function closure_designmat(scanlca::Scan{A,B,C}, scanvis) where {A,B,C<:StructArray{<:EHTLogClosureAmplitudeDatum}}
+    ant1, ant2, ant3, ant4 = baselines(scanlca)
+    antvis1, antvis2 = baselines(scanvis)
+    design_mat = zeros(length(scanlca), length(scanvis))
+    # fill in each row of the design matrix
+    for i in axes(design_mat, 2), j in axes(design_mat,1)
+        a1 = ant1[j]
+        a2 = ant2[j]
+        a3 = ant3[j]
+        a4 = ant4[j]
+
+        av1 = antvis1[i]
+        av2 = antvis2[i]
+        # check leg 1
+        (((av1 == a1)&(av2 == a2)) | (av1 == a2)&(av2 == a1))&& (design_mat[j,i] = 1.0)
+
+        #check leg 2
+        (((av1 == a3)&(av2 == a4)) | (av1 == a4)&(av2 == a3))&& (design_mat[j,i] = 1.0)
+
+        #check leg 3
+        (((av1 == a1)&(av2 == a4)) | (av1 == a4)&(av2 == a1))&& (design_mat[j,i] = -1.0)
+
+        #check leg 4
+        (((av1 == a2)&(av2 == a3)) | (av1 == a3)&(av2 == a2))&& (design_mat[j,i] = -1.0)
+
+    end
+    return design_mat
+end
+
+function minimal_lcamp(obsc; kwargs...)
+
+    obs = obsc.copy()
+    # reorder to maximize the snr
+    obs.reorder_tarr_snr()
+
+    lcamp = extract_lcamp(obs; count="max")
+    stlca = scantable(lcamp)
+
+    #Now make the vis obs
+    dvis = extract_vis(obs)
+    st = scantable(dvis)
+
+    minset = _minimal_closure(stlca, st)
+
+    # now create EHTObservation
+    ra, dec = getradec(obs)
+    mjd = getmjd(obs)
+    source = getsource(obs)
+    bw = obs.bw
+    rf = obs.rf
+    return EHTObservation(data = minset, mjd = mjd,
+                          ra = ra, dec= dec,
+                          bandwidth=bw, frequency=rf,
+                          source = source
[…]
