+++ src/observations/ehtim.jl
@@ -156,9 +156,10 @@ function extract_amp(obs; kwargs...)
     source = getsource(obs)
     bw = obs.bw
     rf = obs.rf
-
+    ac = _arrayconfig(data, bw, rf)
     return Comrade.EHTObservation(data = data, mjd = mjd,
                    ra = ra, dec= dec,
+                   config = ac,
                    bandwidth=bw, frequency=rf,
                    source = source,
     )
@@ -171,15 +172,20 @@ Extracts the complex visibilities from an ehtim observation object
 
 Returns an EHTObservation with complex visibility data
 """
-function extract_vis(obs)
+function extract_vis(obsc; kwargs...)
+    obs = obsc.copy()
+    obs.reorder_tarr_snr()
+
     data = getvisfield(obs)
     ra, dec = getradec(obs)
     mjd = getmjd(obs)
     source = getsource(obs)
     bw = obs.bw
     rf = obs.rf
+    ac = _arrayconfig(data, bw, rf)
     return Comrade.EHTObservation(
                    data = data, mjd = mjd,
+                   config=ac,
                    ra = ra, dec= dec,
                    bandwidth=bw, frequency=rf,
                    source = source,
@@ -191,12 +197,12 @@ function closure_designmat(scancp::Scan{A,B,C}, scanvis) where {A,B,C<:StructArr
     ant1, ant2, ant3 = baselines(scancp)
     antvis1, antvis2 = baselines(scanvis)
     design_mat = zeros(length(scancp), length(scanvis))
+    #throw("here")
     # fill in each row of the design matrix
     for i in axes(design_mat, 2), j in axes(design_mat,1)
         a1 = ant1[j]
         a2 = ant2[j]
         a3 = ant3[j]
-
         # check leg 1
         ((antvis1[i] == a1) & (antvis2[i] == a2)) && (design_mat[j,i] = 1.0)
         ((antvis1[i] == a2) & (antvis2[i] == a1)) && (design_mat[j,i] = -1.0)
@@ -247,14 +253,14 @@ function minimal_lcamp(obsc; kwargs...)
     # reorder to maximize the snr
     obs.reorder_tarr_snr()
 
-    lcamp = extract_lcamp(obs; count="max")
+    lcamp = _ehtim_lcamp(obs; count="max", kwargs...)
     stlca = scantable(lcamp)
 
     #Now make the vis obs
-    dvis = extract_vis(obs)
+    dvis = extract_vis(obsc)
     st = scantable(dvis)
 
-    minset = _minimal_closure(stlca, st)
+    minset, dmat = _minimal_closure(stlca, st)
 
     # now create EHTObservation
     ra, dec = getradec(obs)
@@ -262,13 +268,144 @@ function minimal_lcamp(obsc; kwargs...)
     source = getsource(obs)
     bw = obs.bw
     rf = obs.rf
+    ac = arrayconfig(dvis)
+    clac = ClosureConfig(ac, dmat)
     return EHTObservation(data = minset, mjd = mjd,
+                          config=clac,
                           ra = ra, dec= dec,
                           bandwidth=bw, frequency=rf,
-                          source = source
+                          source = source,
                         )
 
 end
+
+
+function _ehtim_cphase(obsc; count="max", kwargs...)
+    obs = obsc.copy()
+
+    obs.reorder_tarr_snr()
+
+    obs.add_cphase(;count=count, kwargs...)
+    data = getcpfield(obs)
+    ra, dec = getradec(obs)
+    mjd = getmjd(obs)
+    source = getsource(obs)
+    bw = obs.bw
+    rf = obs.rf
+
+
+    cphase = EHTObservation(data = data, mjd = mjd,
+                   config=nothing,
[…]
