### 2022-03-31 bbd028e
added docstrings and fixed pulse stuff
