+++ Project.toml
@@ -57,7 +57,7 @@ ArgCheck = "2"
 AstroTime = "0.6,0.7"
 BasicInterpolators = "0.6"
 ChainRulesCore = "1"
-ComradeBase = "0.3"
+ComradeBase = "^0.3.1"
 DensityInterface = "0.4"
 Distributions = "0.24,0.25"
 DocStringExtensions = "0.6,0.7,0.8"

[…]
