### 2022-03-31 dee608c
added new examples
