### 2022-04-07 #4262 @arfon
@ptiede – thanks for your submission to JOSS. Is this submission associated with another paper in a AAS journal?

Also, could you take a look a [this list of potential reviewers](https://bit.ly/joss-reviewers) and identify a few people who would be good candidates to review this submission? (Or feel free to suggest any individuals you make know of who would be good candidates).
