### 2022-04-07 #4262 @ptiede
@arfon This paper isn't associated with an AAS submission, but the code has been used for EHT results and will be applied to future EHT datasets.

Per reviewers from the spreadsheet I think the following people are probably reasonable:

- ygrange
- garrettj403
- eschnett
- OxCoto

For external people, kazuakiyama is very knowledgeable in this area so they would be a reasonable person to ask.
