### 2022-04-08 #4262 @arfon
:wave: @eschnett @kazuakiyama would you be willing to review this submission for JOSS? [_Comrade: Composable Modeling of Radio Emission _](https://github.com/ptiede/Comrade.jl).
