### 2022-04-08 #4262 @arfon
> I have a conflict of interest with the author Paul Tiede @ptiede. He was at my institution last year.

Got it. Thanks for letting me know about this @eschnett. Are/were you collaborators? If not, if you feel like you're able to provide an impartial review I would be interested in having you review regardless, especially since you have Julia experience.
