### 2022-04-14 9221766
CompatHelper: bump compat for MeasureBase to 0.7, (keep existing compat)
