+++ Project.toml
@@ -65,7 +65,7 @@ ForwardDiff = "0.9, 0.10"
 HypercubeTransform = "^0.2.4"
 KeywordCalls = "0.2"
 MappedArrays = "0.3,0.4"
-MeasureBase = "0.5"
+MeasureBase = "0.5, 0.7"
 MeasureTheory = "0.13"
 NFFT = "0.10, 0.11, 0.12"
 PaddedViews = "0.5"
