+++ src/inference/inference.jl
@@ -3,9 +3,9 @@ using AbstractMCMC
 
 using AbstractDifferentiation
 
-struct IsNested end
+struct IsCube end
 
-struct IsMCMC end
+struct IsFlat end
 
 struct HasDeriv end
 struct NoDeriv end
@@ -24,13 +24,13 @@ function AbstractMCMC.sample(post::Posterior, sampler::S, args...; init_params=n
     return _sample(samplertype(S), post, sampler, args...; init_params=θ0, kwargs...)
 end
 
-function _sample(::IsMCMC, post, sampler, args...; init_params, kwargs...)
+function _sample(::IsFlat, post, sampler, args...; init_params, kwargs...)
     tpost = asflat(post)
     θ0 = inverse(tpost.transform, init_params)
     return sample(tpost, sampler, args...; init_params=θ0, kwargs...)
 end
 
-function _sample(::IsNested, post, sampler, args...; init_params, kwargs...)
+function _sample(::IsCube, post, sampler, args...; init_params, kwargs...)
     tpost = ascube(post)
     θ0 = inverse(tpost.transform, init_params)
     return sample(tpost, sampler, args...; init_params=θ0, kwargs...)

[…]
