+++ Project.toml
@@ -29,6 +29,7 @@ MappedArrays = "dbb5928d-eab1-5f90-85c2-b9b0edb7c900"
 MeasureBase = "fa1605e6-acd5-459c-a1e6-7e635759db14"
 MeasureTheory = "eadaa1a4-d27c-401d-8699-e962e1bbc33b"
 NFFT = "efe261a4-0d2b-5849-be55-fc731d526b0d"
+NamedTupleTools = "d9ec5142-1e00-5aa0-9d6a-321866360f50"
 PaddedViews = "5432bcbf-9aad-5242-b902-cca2824c8663"
 ParameterHandling = "2412ca09-6db7-441c-8e3a-88d5709968c5"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
