### 2022-04-16 9e8ca12
updated hmc to use parallel methods
