+++ src/inference/inference.jl
@@ -17,21 +17,15 @@ export sample
 
 
 function AbstractMCMC.sample(post::Posterior, sampler::S, args...; init_params=nothing, kwargs...) where {S}
-    θ0 = init_params
-    if isnothing(init_params)
-        θ0 = rand(post.prior)
-    end
-    return _sample(samplertype(S), post, sampler, args...; init_params=θ0, kwargs...)
+    return _sample(samplertype(S), post, sampler, args...; init_params, kwargs...)
 end
 
 function _sample(::IsFlat, post, sampler, args...; init_params, kwargs...)
     tpost = asflat(post)
-    θ0 = inverse(tpost.transform, init_params)
-    return sample(tpost, sampler, args...; init_params=θ0, kwargs...)
+    return sample(tpost, sampler, args...; init_params, kwargs...)
 end
 
 function _sample(::IsCube, post, sampler, args...; init_params, kwargs...)
     tpost = ascube(post)
-    θ0 = inverse(tpost.transform, init_params)
-    return sample(tpost, sampler, args...; init_params=θ0, kwargs...)
+    return sample(tpost, sampler, args...; init_params, kwargs...)
 end
