+++ src/inference/advancedhmc.jl
@@ -1,4 +1,5 @@
 using .AdvancedHMC
+using Random
 export  HMC
 
 Base.@kwdef struct HMC{S,I,P,T,A,D}
@@ -34,6 +35,47 @@ function make_pullback(ℓ, autodiff::Function)
     end
 end
 
+
+function _initialize_hmc(tpost::TransformedPosterior, init_params, nchains)
+    isnothing(init_params) && return inverse.(Ref(tpost.transform), rand(tpost.lpost.prior, nchains))
+    @argcheck length(init_params) == nchains
+    return init_params
+end
+
+
+
+function AbstractMCMC.sample(tpost::TransformedPosterior,
+                             sampler::HMC, parallel::AbstractMCMC.AbstractMCMCEnsemble,
+                             nsamples, nchains, args...;
+                             init_params=nothing, kwargs...
+                             )
+
+    ℓ(x) = logdensity(tpost, x)
+
+    ∇ℓ = make_pullback(ℓ, sampler.autodiff)
+    θ0 = _initialize_hmc(tpost, init_params, nchains)
+
+
+    model = AdvancedHMC.DifferentiableDensityModel(ℓ, ∇ℓ)
+    metric = sampler.metric
+    # This is a hack to get a good initial step size
+    hamiltonian = Hamiltonian(metric, ℓ, ∇ℓ)
+    ϵ0 = find_good_stepsize(hamiltonian, first(θ0))
+    integrator = sampler.integrator(ϵ0)
+
+    # form the HMCKernel
+    kernel = HMCKernel(Trajectory{sampler.trajectory}(integrator, sampler.termination))
+    adaptor = sampler.adaptor(MassMatrixAdaptor(metric), StepSizeAdaptor(sampler.targetacc, integrator))
+
+    res = AbstractMCMC.sample(Random.GLOBAL_RNG, model, kernel, metric, adaptor, parallel, nsamples, nchains, args...; init_params=θ0, chain_type=Array, kwargs...)
+
+    stats = [TupleVector(getproperty.(r, :stat)) for r in res]
+    samples = [getproperty.(getproperty.(r, :z), :θ) for r in res]
+    chains = [TupleVector(transform.(Ref(tpost), s)) for s in samples]
+    return chains, stats
+
+end
+
 function AbstractMCMC.sample(tpost::TransformedPosterior, sampler::HMC, nsamples, args...;
                              init_params=nothing,
                              kwargs...)
@@ -63,6 +105,6 @@ function AbstractMCMC.sample(tpost::TransformedPosterior, sampler::HMC, nsamples
     stats = TupleVector(getproperty.(res, :stat))
     samples = getproperty.(getproperty.(res, :z), :θ)
     chain = transform.(Ref(tpost), samples)
-    return TupleVector(chain), stats
+    return TupleVector(chain), stats, res
     #return TupleVector(transform.(Ref(tpost), chain)), TupleVector(stats)
 end
