### 2022-04-17 6516a3e
CompatHelper: bump compat for MeasureTheory to 0.15, (keep existing compat)
