+++ Project.toml
@@ -66,7 +66,7 @@ HypercubeTransform = "^0.2.4"
 KeywordCalls = "0.2"
 MappedArrays = "0.3,0.4"
 MeasureBase = "0.5"
-MeasureTheory = "0.13"
+MeasureTheory = "0.13, 0.15"
 NFFT = "0.10, 0.11, 0.12"
 PaddedViews = "0.5"
 ParameterHandling = "0.4"
