### 2022-04-19 #4262 @kazuakiyama
Thanks a lot for contacting me, and sorry that my reply was extremely late. I was in vacation in that week, and I just found this request.

I will likely have a conflict of interest with Paul as I’m too close to Paul — we have been working together in the same research collaboration (EHT Collaboration), and I’m affiliated with his current institution (Harvard Univ. Black Hole Initiative). As Paul already mentioned to you, I think more general astronomy developers outside of the EHT collaboration would give the best fit to Comrade.jl.

- Kazu

> On Apr 8, 2022, at 1:57 AM, Arfon Smith ***@***.***> wrote:
>
>
> 👋 @eschnett <https://github.com/eschnett> @kazuakiyama <https://github.com/kazuakiyama> would you be willing to review this submission for JOSS? _Comrade: Composable Modeling of Radio Emission _ <https://github.com/ptiede/Comrade.jl>.
>
> We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html>
