### 2022-04-24 #4262 @arfon
OK thanks @ptiede. I'm asking around a bit in my network to see if people know any Julia-proficient radio astronomers.
