### 2022-04-24 #4262 @arfon
:wave: @mileslucas & @ericagol – would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The software under consideration is a Julia package: _Comrade: Composable Modeling of Radio Emission_
