### 2022-04-25 #4262 @mileslucas
@arfon Sure, I'm happy to review this project, although my radio astronomy knowledge is limited.
