### 2022-05-10 44dda1c
CompatHelper: bump compat for MeasureBase to 0.8, (keep existing compat)
