+++ examples/imaging_closures.jl
@@ -21,10 +21,10 @@ obs = ehtim.obsdata.load_uvfits(joinpath(@__DIR__, "SR1_M87_2017_096_hi_hops_net
 obs.add_scans()
 # kill 0-baselines since we don't care about
 # large scale flux and make scan-average data
-obs = obs.flag_uvdist(uv_min=0.1e9).avg_coherent(0.0, scan_avg=true).add_fractional_noise(0.01)
+obs = obs.avg_coherent(0.0, scan_avg=true).add_fractional_noise(0.01)
 # extract log closure amplitudes and closure phases
 dlcamp = extract_lcamp(obs)
-dcphase = extract_cphase(obs)
+dcphase = extract_cphase(obs; cut_trivial=true)
 lklhd = RadioLikelihood(dlcamp, dcphase)
 
 # build the model here we fit a ring with a azimuthal
@@ -41,16 +41,17 @@ function (model::Model)(θ)
     c  = θ.c
     cmat = reshape(c, model.npix, model.npix)
     img = IntensityMap(cmat, model.fov, model.fov, BSplinePulse{3}())
-    return modelimage(img, model.cache)
+    return modelimage(img, model.cache) + stretched(Gaussian(), μas2rad(250.0), μas2rad(250.0))
 end
 
-fovxy = μas2rad(65.0)
-npix = 8
+fovxy = μas2rad(80.0)
+npix = 15
 prior = (
-          c = DistributionsAD.TuringDirichlet(npix*npix, 0.5),
-          #c = MvLogNormal(fill(-4.0, npix*npix), 2.0),
+          #c = DistributionsAD.TuringDirichlet(npix*npix, 0.5),
+          c = MvLogNormal(fill(log(inv(4*npix^2)), npix*npix), 3.0),
         )
 # Now form the posterior
+u, v = getuv(dlcamp.config)
 cache = create_cache(Comrade.DFTAlg(dlcamp), IntensityMap(rand(npix,npix), fovxy, fovxy, BSplinePulse{3}()))
 mms = Model(cache, fovxy, npix)
 
@@ -60,7 +61,7 @@ tpost = asflat(post)
 # First to reduce burn in we use pathfinder
 ndim = dimension(tpost)
 f, tr = GalacticOptim.OptimizationFunction(post, GalacticOptim.AutoZygote())
-x0 = inverse(tr, rand(post.prior))
+x0 = HypercubeTransform.inverse(tr, rand(post.prior))
 prob = GalacticOptim.OptimizationProblem(f, x0, nothing, lb=fill(-5.0, dimension(tr)), ub=fill(5.0, dimension(tr)))
 sol, xopt = solve(prob, BBO_adaptive_de_rand_1_bin_radiuslimited(), tr; maxiters=1000_000)
 
@@ -68,7 +69,7 @@ sol, xopt = solve(prob, BBO_adaptive_de_rand_1_bin_radiuslimited(), tr; maxiters
 
 prob = GalacticOptim.OptimizationProblem(f, randn(ndim), nothing)
 sol, xopt = solve(prob, LBFGS(), tr; show_trace=true, show_every=50, iterations=4000, g_tol=1e-2)
-
+@info sol.minimum
 
 residual(mms(xopt), dlcamp)
 residual(mms(xopt), dcphase)
@@ -78,8 +79,8 @@ using AdaptiveMCMC
 cacc, sacc = sample(post, Comrade.AdaptMCMC(ntemp=5), 500_000, 250_000; init_params=xopt, thin=50)
 
 # now we sample using hmc
-metric = DiagEuclideanMetric(ndim)
[…]
