### 2022-05-21 0374e9e
CompatHelper: add new compat entry for NamedTupleTools at version 0.14, (keep existing compat)
