+++ Project.toml
@@ -72,6 +72,7 @@ MappedArrays = "0.3,0.4"
 MeasureBase = "0.7"
 MeasureTheory = "0.15"
 NFFT = "0.10, 0.11, 0.12"
+NamedTupleTools = "0.14"
 PaddedViews = "0.5"
 ParameterHandling = "0.4"
 PyCall = "^1"
