+++ lib/ComradeAHMC/test/runtests.jl
@@ -1,0 +1,6 @@
+using Comrade, ComradeAHMC
+using Test
+
+@testset "ComradeAHMC.jl" begin
+
+end
