+++ src/inference/advancedhmc.jl
@@ -1,3 +1,0 @@
-using .AdvancedHMC
-using Random
-export  HMC
