+++ lib/ComradeAdaptMCMC/test/runtests.jl
@@ -1,0 +1,6 @@
+using Comrade, ComradeAdaptMCMC
+using Test
+
+@testset "ComradeAdaptMCMC.jl" begin
+
+end
