+++ Project.toml
@@ -1,6 +1,6 @@
 name = "Comrade"
 uuid = "99d987ce-9a1e-4df8-bc0b-1ea019aa547b"
-authors = ["Paul Tiede"]
+authors = ["Paul Tiede <ptiede91@gmail.com>"]
 version = "0.3.1"
 
 [deps]
