+++ lib/ComradeDynesty/test/runtests.jl
@@ -1,0 +1,6 @@
+using Comrade, ComradeDynesty
+using Test
+
+@testset "ComradeDynesty.jl" begin
+
+end
