+++ lib/ComradeDynesty/Project.toml
@@ -1,0 +1,21 @@
+name = "ComradeDynesty"
+uuid = "0e22d245-d796-46d0-bd67-132aa2ceb697"
+authors = ["Paul Tiede <ptiede91@gmail.com>"]
+version = "0.1.0"
+
+[deps]
+Comrade = "99d987ce-9a1e-4df8-bc0b-1ea019aa547b"
+Dynesty = "eb527566-0f3e-4aab-bb5f-9d2e403dba70"
+TupleVectors = "615932cf-77b6-4358-adcd-5b7eba981d7e"
+
+[compat]
+Comrade = "0.3"
+Dynesty = "0.1"
+TupleVectors = "0.1"
+
+[extras]
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[targets]
+test = ["Test"]
+
