### 2022-05-21 f6b66c8
removed galatic from main
