+++ src/inference/galacticoptim.jl
@@ -1,1 +1,0 @@
-using .GalacticOptim
