+++ lib/ComradeNested/Project.toml
@@ -1,0 +1,21 @@
+name = "ComradeNested"
+uuid = "233675ba-9426-43ce-b06d-28ea4529d7f1"
+authors = ["Paul Tiede <ptiede91@gmail.com>"]
+version = "0.1.0"
+
+[deps]
+Comrade = "99d987ce-9a1e-4df8-bc0b-1ea019aa547b"
+NestedSamplers = "41ceaf6f-1696-4a54-9b49-2e7a9ec3782e"
+TupleVectors = "615932cf-77b6-4358-adcd-5b7eba981d7e"
+
+[compat]
+Comrade = "0.3"
+NestedSamplers = "0.8"
+TupleVectors = "0.1"
+
+[extras]
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[targets]
+test = ["Test"]
+
