### 2022-05-21 fff3aaa
add prior_sample
