+++ lib/ComradeNested/Project.toml
@@ -4,8 +4,10 @@ authors = ["Paul Tiede <ptiede91@gmail.com>"]
 version = "0.1.0"
 
 [deps]
+AbstractMCMC = "80f14c24-f653-4e6a-9b94-39d6b0f70001"
 Comrade = "99d987ce-9a1e-4df8-bc0b-1ea019aa547b"
 NestedSamplers = "41ceaf6f-1696-4a54-9b49-2e7a9ec3782e"
+Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 TupleVectors = "615932cf-77b6-4358-adcd-5b7eba981d7e"
 
 [compat]
@@ -18,4 +20,3 @@ Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
 test = ["Test"]
-
