+++ lib/ComradeAdaptMCMC/Project.toml
@@ -4,6 +4,7 @@ authors = ["Paul Tiede <ptiede91@gmail.com>"]
 version = "0.1.0"
 
 [deps]
+AbstractMCMC = "80f14c24-f653-4e6a-9b94-39d6b0f70001"
 AdaptiveMCMC = "717c3277-546c-407d-8270-09a39a0919a0"
 Comrade = "99d987ce-9a1e-4df8-bc0b-1ea019aa547b"
 TupleVectors = "615932cf-77b6-4358-adcd-5b7eba981d7e"
@@ -18,4 +19,3 @@ Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
 test = ["Test"]
-
