+++ lib/ComradeGalactic/Project.toml
@@ -6,6 +6,7 @@ version = "0.1.0"
 [deps]
 Comrade = "99d987ce-9a1e-4df8-bc0b-1ea019aa547b"
 GalacticOptim = "a75be94c-b780-496d-a8a9-0878b188d577"
+Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 
 [compat]
 Comrade = "0.3"
