+++ Project.toml
@@ -70,7 +70,7 @@ HypercubeTransform = "^0.2.7"
 KeywordCalls = "0.2"
 MappedArrays = "0.3,0.4"
 MeasureBase = "0.7"
-MeasureTheory = "0.15"
+MeasureTheory = "0.15, 0.16"
 NFFT = "0.10, 0.11, 0.12"
 PaddedViews = "0.5"
 ParameterHandling = "0.4"
