### 2022-05-27 d998fd5
CompatHelper: bump compat for DocStringExtensions to 0.9, (keep existing compat)
