+++ Project.toml
@@ -61,7 +61,7 @@ ChainRulesCore = "1"
 ComradeBase = "^0.3.1"
 DensityInterface = "0.4"
 Distributions = "0.24,0.25"
-DocStringExtensions = "0.6,0.7,0.8"
+DocStringExtensions = "0.6,0.7,0.8, 0.9"
 FFTW = "^1"
 FLoops = "0.2"
 FillArrays = "0.12, 0.13"
