+++ src/distributions/radiolikelihood.jl
@@ -1,10 +1,20 @@
 using MeasureTheory
-export RadioLikelihood, logdensity
+export RadioLikelihood, logdensityof, MultiRadioLikelihood
 using LinearAlgebra
 
 struct RadioLikelihood{T,A} <: MeasureBase.AbstractMeasure
     lklhds::T
     ac::A
+end
+
+struct MultiRadioLikelihood{L} <: MeasureBase.AbstractMeasure
+    lklhds::L
+end
+
+MultiRadioLikelihood(lklhds::RadioLikelihood...) = MultiRadioLikelihood(lklhds)
+
+function logdensityof(lklhds::MultiRadioLikelihood, m)
+    sum(x->logdensityof(l, m), lklhds)
 end
 
 function RadioLikelihood(data::EHTObservation...)
@@ -32,13 +42,13 @@ phase(vis::AbstractArray{<:Complex}) = angle.(vis)
 function MeasureBase.logdensityof(d::RadioLikelihood, m::ComradeBase.AbstractModel)
     ac = d.ac
     vis = visibilities(m, ac)
-    return MeasureBase.logdensityof(d, vis)
+    return logdensityof(d, vis)
 end
 
 function MeasureBase.logdensityof(d::RadioLikelihood, vis::AbstractArray)
     # We use a for loop here since Zygote plays nice with this
     acc = logdensityof(first(d.lklhds), vis)
-    @inbounds for l in d.lklhds[begin+1:end]
+    @inbounds for l in @view(d.lklhds[begin+1:end])
         acc += logdensityof(l, vis)
     end
     return acc
