+++ lib/ComradeGalactic/Project.toml
@@ -9,7 +9,7 @@ GalacticOptim = "a75be94c-b780-496d-a8a9-0878b188d577"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 
 [compat]
-Comrade = "0.3"
+Comrade = "0.3,0.4"
 GalacticOptim = "3"
 
 [extras]
