+++ lib/ComradeAdaptMCMC/Project.toml
@@ -11,7 +11,7 @@ TupleVectors = "615932cf-77b6-4358-adcd-5b7eba981d7e"
 
 [compat]
 AdaptiveMCMC = "0.1"
-Comrade = "0.3"
+Comrade = "0.3,0.4"
 TupleVectors = "0.1"
 
 [extras]
