+++ .github/workflows/ci.yml
@@ -17,9 +17,7 @@ jobs:
           - ComradeDynesty
           - ComradeGalactic
           - ComradeNested
-          - ComradePathfinder
         version:
-          - '1.6'
           - '1.7'
         os:
           - ubuntu-latest

[…]
