### 2022-05-29 190b59f
added dynesty tests
