+++ lib/ComradeDynesty/Project.toml
@@ -6,6 +6,7 @@ version = "0.1.0"
 [deps]
 AbstractMCMC = "80f14c24-f653-4e6a-9b94-39d6b0f70001"
 Comrade = "99d987ce-9a1e-4df8-bc0b-1ea019aa547b"
+Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 Dynesty = "eb527566-0f3e-4aab-bb5f-9d2e403dba70"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 TupleVectors = "615932cf-77b6-4358-adcd-5b7eba981d7e"
