+++ test/Project.toml
@@ -1,15 +1,11 @@
 [deps]
-AdvancedHMC = "0bf59076-c3b1-5ca4-86bd-e02cd72cde3d"
-BlackBoxOptim = "a134a8b2-14d6-55f6-9291-3336d3ab0209"
-ComradeAdaptMCMC = "5a6518a2-e6ab-4020-97b8-110ebf16b849"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
+GalacticBBO = "80c49c3a-6557-47d9-8f5b-13d0a2920315"
 GalacticOptim = "a75be94c-b780-496d-a8a9-0878b188d577"
 MeasureBase = "fa1605e6-acd5-459c-a1e6-7e635759db14"
-NestedSamplers = "41ceaf6f-1696-4a54-9b49-2e7a9ec3782e"
 Optim = "429524aa-4258-5aef-a3af-852621145aeb"
-Pathfinder = "b1d3bc72-d0e7-4279-b92f-7fa5d6d2d454"
 Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 PyCall = "438e738f-606a-5dbb-bf0a-cddfbfd45ab0"
