+++ lib/ComradeAHMC/test/runtests.jl
@@ -1,6 +1,7 @@
 using Comrade, ComradeAHMC, Distributions
 using Test
 
+load_ehtim()
 include(joinpath(@__DIR__, "../../../test/test_util.jl"))
 
 @testset "ComradeAHMC.jl" begin
