+++ lib/ComradeAHMC/Project.toml
@@ -9,16 +9,17 @@ AbstractMCMC = "80f14c24-f653-4e6a-9b94-39d6b0f70001"
 AdvancedHMC = "0bf59076-c3b1-5ca4-86bd-e02cd72cde3d"
 ArgCheck = "dce04be8-c92d-5529-be00-80e4d2c0e197"
 Comrade = "99d987ce-9a1e-4df8-bc0b-1ea019aa547b"
-Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
-HypercubeTransform = "9ec9aee3-0fd3-44c2-8e61-a50acc66f3c8"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 TupleVectors = "615932cf-77b6-4358-adcd-5b7eba981d7e"
 
 [compat]
+AbstractDifferentiation = "0.4"
 AbstractMCMC = "3,4"
 AdvancedHMC = "0.3"
+ArgCheck = "2"
 Comrade = "0.3,0.4"
+Reexport = "1"
 TupleVectors = "0.1"
 julia = "1.6"
 
