### 2022-05-29 8028cde
CompatHelper: add new compat entry for Plots at version 1 for package docs, (keep existing compat)
