### 2022-05-29 df8a8a8
CompatHelper: add new compat entry for Pathfinder at version 0.4 for package docs, (keep existing compat)
