### 2022-05-29 f99b47b
CompatHelper: add new compat entry for AdvancedHMC at version 0.3 for package docs, (keep existing compat)
