### 2022-05-29 53ca798
CompatHelper: add new compat entry for Glob at version 1 for package docs, (keep existing compat)
