### 2022-05-29 0f3174f
CompatHelper: add new compat entry for GalacticOptim at version 3 for package docs, (keep existing compat)
