### 2022-05-29 28af4c4
CompatHelper: add new compat entry for Literate at version 2 for package docs, (keep existing compat)
