### 2022-05-29 36695b3
CompatHelper: add new compat entry for ComradeBase at version 0.3 for package docs, (keep existing compat)
