### 2022-05-29 713d718
CompatHelper: add new compat entry for Documenter at version 0.27 for package docs, (keep existing compat)
