### 2022-05-29 1dab913
CompatHelper: add new compat entry for Comrade at version 0.4 for package docs, (keep existing compat)
