### 2022-05-29 9635d3e
CompatHelper: add new compat entry for HypercubeTransform at version 0.2 for package docs, (keep existing compat)
