+++ lib/ComradeGalactic/Project.toml
@@ -11,6 +11,8 @@ Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 [compat]
 Comrade = "0.3,0.4"
 GalacticOptim = "3"
+Reexport = "1"
+julia = "1.6"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
