### 2022-05-30 cddb292
CompatHelper: add new compat entry for Comrade at version 0.4 for package docs, (keep existing compat)
