+++ docs/Project.toml
@@ -1,6 +1,5 @@
 [deps]
 AdvancedHMC = "0bf59076-c3b1-5ca4-86bd-e02cd72cde3d"
-Comrade = "99d987ce-9a1e-4df8-bc0b-1ea019aa547b"
 ComradeBase = "6d8c423b-a35f-4ef1-850c-862fe21f82c4"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
