### 2022-05-30 9e142d9
CompatHelper: add new compat entry for Glob at version 1 for package docs, (keep existing compat)
