### 2022-05-30 874e7ad
CompatHelper: add new compat entry for Documenter at version 0.27 for package docs, (keep existing compat)
