### 2022-05-30 a3ff83f
CompatHelper: add new compat entry for Literate at version 2 for package docs, (keep existing compat)
