+++ docs/Project.toml
@@ -13,3 +13,4 @@ Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 [compat]
 Distributions = "0.25"
 HypercubeTransform = "0.2"
+Literate = "2"
