### 2022-05-30 f294629
CompatHelper: add new compat entry for Plots at version 1 for package docs, (keep existing compat)
