### 2022-05-30 f0399f2
CompatHelper: add new compat entry for AdvancedHMC at version 0.3 for package docs, (keep existing compat)
