### 2022-05-30 e5d6746
CompatHelper: add new compat entry for GalacticOptim at version 3 for package docs, (keep existing compat)
