+++ docs/Project.toml
@@ -12,4 +12,5 @@ Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 
 [compat]
 Distributions = "0.25"
+GalacticOptim = "3"
 HypercubeTransform = "0.2"
