### 2022-05-30 a227b15
CompatHelper: add new compat entry for ComradeBase at version 0.3 for package docs, (keep existing compat)
