+++ docs/Project.toml
@@ -11,5 +11,6 @@ Pathfinder = "b1d3bc72-d0e7-4279-b92f-7fa5d6d2d454"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 
 [compat]
+ComradeBase = "0.3"
 Distributions = "0.25"
 HypercubeTransform = "0.2"
