### 2022-05-30 ea914c7
CompatHelper: add new compat entry for Pathfinder at version 0.4 for package docs, (keep existing compat)
