### 2022-05-30 4a185cd
remove printing in tests
