### 2022-05-30 505e7f1
add ComradeGalactic tests
