+++ .github/workflows/ci.yml
@@ -15,7 +15,7 @@ jobs:
           - ComradeAdaptMCMC
           - ComradeAHMC
           - ComradeDynesty
-          #- ComradeGalactic
+          - ComradeGalactic
           - ComradeNested
         version:
           - '1.7'
