### 2022-05-30 56aaac0
fix galactic test
