### 2022-05-30 c305779
fix galactic extra
