+++ test/Core/core.jl
@@ -8,4 +8,5 @@ using Plots
 include(joinpath(@__DIR__, "observation.jl"))
 include(joinpath(@__DIR__, "distributions.jl"))
 include(joinpath(@__DIR__, "models.jl"))
+include(joinpath(@__DIR__, "gains.jl"))
 include(joinpath(@__DIR__, "bayes.jl"))
