+++ docs/Project.toml
@@ -1,13 +1,14 @@
 [deps]
-AdvancedHMC = "0bf59076-c3b1-5ca4-86bd-e02cd72cde3d"
+ComradeAHMC = "a4336a5c-78bc-4363-8a90-ce3fa9d3abe4"
 ComradeBase = "6d8c423b-a35f-4ef1-850c-862fe21f82c4"
+ComradeGalactic = "47701775-187b-49a4-b924-794639d71cdc"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+GalacticBBO = "80c49c3a-6557-47d9-8f5b-13d0a2920315"
 GalacticOptim = "a75be94c-b780-496d-a8a9-0878b188d577"
 Glob = "c27321d9-0574-5035-807b-f59d2c89b15c"
 HypercubeTransform = "9ec9aee3-0fd3-44c2-8e61-a50acc66f3c8"
 Literate = "98b081ad-f1c9-55d3-8b20-4c87d4299306"
-Pathfinder = "b1d3bc72-d0e7-4279-b92f-7fa5d6d2d454"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 
 [compat]

[…]
