### 2022-06-01 b3f9039
CompatHelper: add new compat entry for FITSIO at version 0.16, (keep existing compat)
