+++ Project.toml
@@ -67,6 +67,7 @@ DensityInterface = "0.4"
 Distributions = "0.24,0.25"
 DocStringExtensions = "0.6,0.7,0.8, 0.9"
 FFTW = "^1"
+FITSIO = "0.16"
 FLoops = "0.2"
 FillArrays = "0.12, 0.13"
 ForwardDiff = "0.9, 0.10"
