### 2022-06-01 7e356a7
CompatHelper: add new compat entry for FileIO at version 1, (keep existing compat)
