+++ Project.toml
@@ -68,6 +68,7 @@ Distributions = "0.24,0.25"
 DocStringExtensions = "0.6,0.7,0.8, 0.9"
 FFTW = "^1"
 FLoops = "0.2"
+FileIO = "1"
 FillArrays = "0.12, 0.13"
 ForwardDiff = "0.9, 0.10"
 HypercubeTransform = "^0.2.7"
