### 2022-06-01 afe9813
CompatHelper: add new compat entry for GalacticBBO at version 0.1 for package docs, (keep existing compat)
