+++ docs/Project.toml
@@ -13,4 +13,5 @@ Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 
 [compat]
 Distributions = "0.25"
+GalacticBBO = "0.1"
 HypercubeTransform = "0.2"
