+++ test/Core/core.jl
@@ -5,8 +5,11 @@ using Statistics
 using Test
 using Plots
 
+include(joinpath(@__DIR__, "../test_util.jl"))
+
 include(joinpath(@__DIR__, "observation.jl"))
 include(joinpath(@__DIR__, "distributions.jl"))
 include(joinpath(@__DIR__, "models.jl"))
 include(joinpath(@__DIR__, "gains.jl"))
+include(joinpath(@__DIR__, "io.jl"))
 include(joinpath(@__DIR__, "bayes.jl"))
