+++ examples/paper_example.jl
@@ -1,10 +1,10 @@
 using Pkg; Pkg.activate(@__DIR__)
 using Comrade
 using Distributions
-using Pathfinder
-using AdvancedHMC
+using ComradeGalactic
+using ComradeAHMC
+using GalacticBBO
 using Plots
-using TupleVectors
 using StatsBase
 
 # load eht-imaging we use this to load eht data
@@ -17,7 +17,7 @@ obs.add_scans()
 obs = obs.flag_uvdist(uv_min=0.1e9).avg_coherent(0.0, scan_avg=true)
 # extract log closure amplitudes and closure phases
 dlcamp = extract_lcamp(obs)
-dcphase = extract_cphase(obs; cutmin)
+dcphase = extract_cphase(obs)
 # form the likelihood
 lklhd = RadioLikelihood(dlcamp, dcphase)
 # build the model here we fit a ring with a azimuthal
@@ -35,7 +35,7 @@ prior = (
           a = Uniform(-0.5, 0.5), b = Uniform(-0.5, 0.5),
           f = Uniform(0.0, 1.0),
           sig = Uniform(μas2rad(1.0), μas2rad(40.0)),
-          asy = Uniform(0.0, 0.75),
+          asy = Uniform(0.0, 0.9),
           pa = Uniform(0.0, 1π),
           x = Uniform(-μas2rad(80.0), μas2rad(80.0)),
           y = Uniform(-μas2rad(80.0), μas2rad(80.0))
@@ -43,17 +43,27 @@ prior = (
 # Now form the posterior
 post = Posterior(lklhd, prior, model)
 # We will use HMC to sample the posterior.
-# First to reduce burn in we use pathfinder
-using Dynesty
-chain, stats = sample(post, NestedSampler(dimension(post)))
-echain = sample(chain, Weights(stats.weights), 10_000)|> TupleVector
+# First we will find a reasonable starting location using GalacticOptim
+# For optimization we need to specify what transform to use. Here we will transform to
+# the unit hypercube
+tpost = ascube(post)
+ndim = dimension(tpost)
+f = OptimizationFunction(tpost)
+prob = OptimizationProblem(f, rand(ndim), nothing, lb=fill(0.01, ndim), ub = fill(0.99, ndim))
+sol = solve(prob, BBO_adaptive_de_rand_1_bin_radiuslimited(); maxiters=50_000)
 
-residual(model(chain[end]), dlcamp)
-plot(model(echain[rand(eachindex(echain))]), xlims=(-50.0,50.0), ylims=(-50.0,50.0))
+# transform the solution back to regular space
+xopt = transform(tpost, sol)
+# Let's see if the best fit looks reasonable by plotting normalized residuals for the
+# log-closure amplitudes
+residual(model(xopt), dlcamp)
+# we can also plot the best fit model or maximum likelihood estimate (MLE)
+plot(model(xopt), xlims=(-80.0,80.0),ylims=(-80.0,80.0), colorbar=nothing, title="MLE M87")
 
-q, phi, _ = multipathfinder(post, 100)
-# now we sample using hmc
-metric = DiagEuclideanMetric(dimension(post))
-chain, stats = sample(post, HMC(;metric), 2000; nadapts=1000, init_params=chain[end])
-# plot a draw from the posterior
-plot(model(chain[rand(1:length(chain))]), xlims=(-80.0, 80.0), ylims=(-80.0,80.0), colorbar=nothing)
+# Comrade is all about uncertainty quantification so now let's find the posterior!
+# To do this we will use the `AdvancedHMC` package or rather its interface to Comrade.
+metric = DiagEuclideanMetric(ndim)
[…]
