+++ examples/imaging_closures.jl
@@ -1,17 +1,12 @@
 using Pkg; Pkg.activate(@__DIR__)
 using Comrade
 using Distributions
-using Pathfinder
-using AdvancedHMC
+using ComradeGalactic
+using ComradeAHMC
+using GalacticBBO
 using Plots
-using Zygote
-using Parameters
-using GalacticOptim
-using Optim
-using NLopt
-using BlackBoxOptim
-using Dynesty
-using AdaptiveMCMC
+using StatsBase
+using GalacticOptimJL
 using DistributionsAD
 
 # load eht-imaging we use this to load eht data
@@ -21,69 +16,64 @@ obs = ehtim.obsdata.load_uvfits(joinpath(@__DIR__, "SR1_M87_2017_096_hi_hops_net
 obs.add_scans()
 # kill 0-baselines since we don't care about
 # large scale flux and make scan-average data
-obs = obs.avg_coherent(0.0, scan_avg=true).add_fractional_noise(0.01)
+obs = obs.flag_uvdist(uv_min=0.1e9).avg_coherent(0.0, scan_avg=true).add_fractional_noise(0.02)
 # extract log closure amplitudes and closure phases
 dlcamp = extract_lcamp(obs)
-dcphase = extract_cphase(obs; cut_trivial=true)
+dcphase = extract_cphase(obs)
 lklhd = RadioLikelihood(dlcamp, dcphase)
 
-# build the model here we fit a ring with a azimuthal
-#brightness variation and a Gaussian
-
-struct Model{C,F}
+# Build the Model. Here we we a struct to hold some caches
+# which will speed up imaging
+struct ImModel{C}
     cache::C
-    fov::F
+    fov::Float64
     npix::Int
 end
 
-
-function (model::Model)(θ)
-    c  = θ.c
+# For our model we will be using a rasterized image. This can be viewed as something like a
+# non-parametric model. As a result of this we will need to use a `modelimage` object to
+# store cache information we will need to compute the numerical FT.
+function (model::ImModel)(θ)
+    (;c) = θ
+    #Construct the image model
     cmat = reshape(c, model.npix, model.npix)
     img = IntensityMap(cmat, model.fov, model.fov, BSplinePulse{3}())
-    return modelimage(img, model.cache) + stretched(Gaussian(), μas2rad(250.0), μas2rad(250.0))
+    #Create the modelimage object that will use a cache to compute the DFT
+    m = modelimage(img, cache)
 end
-
-fovxy = μas2rad(80.0)
-npix = 15
-prior = (
-          #c = DistributionsAD.TuringDirichlet(npix*npix, 0.5),
-          c = MvLogNormal(fill(log(inv(4*npix^2)), npix*npix), 3.0),
-        )
-# Now form the posterior
-u, v = getuv(dlcamp.config)
+const npix = 10
+const fovxy = μas2rad(65.0)
+# Now we can feed in the array information to form the cache. We will be using a DFT since
+# it is efficient for so few pixels
 cache = create_cache(Comrade.DFTAlg(dlcamp), IntensityMap(rand(npix,npix), fovxy, fovxy, BSplinePulse{3}()))
-mms = Model(cache, fovxy, npix)
+mms = ImModel(cache, fovxy, npix)
+# We will use a Dirichlet prior to enforce that the flux sums to unity since closures are
+# degenerate to total flux.
+prior = (c = DistributionsAD.TuringDirichlet(npix*npix, 1.0),)
 
 post = Posterior(lklhd, prior, mms)
 tpost = asflat(post)
[…]
