### 2022-06-02 d88ea97
CompatHelper: add new compat entry for ComradeAHMC at version 0.1 for package docs, (keep existing compat)
