+++ docs/Project.toml
@@ -12,5 +12,6 @@ Literate = "98b081ad-f1c9-55d3-8b20-4c87d4299306"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 
 [compat]
+ComradeAHMC = "0.1"
 Distributions = "0.25"
 HypercubeTransform = "0.2"
