+++ docs/src/examples/black_hole_image.jl
@@ -12,8 +12,10 @@
 using Comrade
 using Plots
 using Distributions
-using Pathfinder
-using AdvancedHMC
+using GalacticOptim
+using GalacticBBO
+using ComradeGalactic
+using ComradeAHMC
 
 # Pathfinder and AdvancedHMC will be required for sampling the posterior.
 # Distributions is a general purpose probability distribution package in Julia.
@@ -28,11 +30,11 @@ obs.add_scans()
 # kill 0-baselines since we don't care about large scale flux and make scan-average data
 obs = obs.flag_uvdist(uv_min=0.1e9).avg_coherent(0.0, scan_avg=true)
 # grab data products we want to fit: log closure amplitudes and closure phases
-dlcamp = extract_lcamp(obs; count="min")
-dcphase = extract_cphase(obs, count="min")
+dlcamp = extract_lcamp(obs)
+dcphase = extract_cphase(obs)
 
 # For this demo only consider closure products since these are invariant to station specific
-# gain systematics. Given these data products we can then form our visibility likelihood:
+# gain systematics. Given these data products we can then form our radio likelihood:
 
 lklhd = RadioLikelihood(dlcamp, dcphase)
 
@@ -44,8 +46,8 @@ lklhd = RadioLikelihood(dlcamp, dcphase)
 # For the image model we will be using a modified `MRing`, which is a
 # infinitely thin delta ring with an azimuthal structure given by a Fourier expansion.
 # To give the ring some width we will convolve the ring with a gaussian, and add an
-# additional gaussian to the image to model any non-ring flux. The resulting model function
-# is given by
+# additional gaussian to the image to model any non-ring flux. For the model a user
+# must give a function that accepts a named tuple and return the constructed model:
 
 function model(θ)
   (;radius, width, α, β, f, σG, τG, ξG, xG, yG) = θ
@@ -55,7 +57,7 @@ function model(θ)
 end
 
 # We now need to specify the priors for our model. The easiest way to do this is to
-# specify a NamedTuple of distributions.
+# specify a NamedTuple of distributions:
 
 prior = (
           radius = Uniform(μas2rad(10.0), μas2rad(30.0)),
@@ -74,7 +76,7 @@ prior = (
 
 post = Posterior(lklhd, prior, model)
 
-# This constructs a posterior density that can be evaluated by calling `logdensity`.
+# This constructs a posterior density that can be evaluated by calling `logdensityof`.
 # For example
 
 logdensityof(post, (radius = μas2rad(20.0),
@@ -91,51 +93,71 @@ logdensityof(post, (radius = μas2rad(20.0),
 # We can now try to sample from our posterior `post` so that we can make probabilstic
 # inferences about our data.
 
+# Now this model is in **parameter** space. Often optimization and sampling algorithms
+# want it in some modified space. For example, nested sampling algorithms want the
+# parameters in the unit hypercube. To transform the posterior to the unit hypercube we
+# can use the `ascube` function
+
+cpost = ascube(post)
+
+# If we want to flatten the parameter space and move to (-∞, ∞) support we can use the
+# `asflat` function
+
+fpost = asflat(post)
+
+# These transformed posterior expect a vector of parameters. That is we can evaluate the
+# transformed log density by calling
+
+logdensityof(cpost, rand(dimension(cpost)))
+logdensityof(fpost, randn(dimension(fpost)))
+
+# Note that this automatically takes care of the jacobian in the parameter transformation.
+
+# ## Sampling the posterior
+
[…]
