### 2022-06-03 d43c0fc
CompatHelper: add new compat entry for ForwardDiff at version 0.10 for package docs, (keep existing compat)
