+++ docs/Project.toml
@@ -17,4 +17,5 @@ Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 
 [compat]
 Distributions = "0.25"
+ForwardDiff = "0.10"
 HypercubeTransform = "0.2"
