### 2022-06-03 0b4a65b
CompatHelper: add new compat entry for GalacticOptimJL at version 0.1 for package docs, (keep existing compat)
