### 2022-06-03 1037ac3
add docs compat
