### 2022-06-06 #4262 @arfon
:wave: @tammojan – would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The software under consideration is a Julia package: [_Comrade: Composable Modeling of Radio Emission_](https://github.com/ptiede/Comrade.jl)
