### 2022-06-07 #4262 @arfon
> @arfon This looks like an interesting article and software! Could you give an estimate of roughly how much time a reviewer is expected to invest? And what would be the timeline for a review?

Based on my own experience, the initial review (working through a checklist) takes a few hours, and then depending upon the state of the software (and whether additional updates are required by the authors), another hour or two might be required.
