### 2022-06-07 #4262 @arfon
@tammojan – if you decide you're willing and able to help with the review here I'd be delighted to have you as a third reviewer on this submission (as a past JOSS reviewer, @cescalara knows what they are signing up for 😄).
