### 2022-06-07 #4457 @editorialbot
[REVIEW]: Comrade: Composable Modeling of Radio Emission
**Submitting author:** @ptiede (<a href="http://orcid.org/0000-0003-3826-5648">Paul Tiede</a>)
**Repository:** https://github.com/ptiede/Comrade.jl
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.3.1
