### 2022-06-12 c39de25
clean up some gain stuff
