+++ examples/gain_m87.jl
@@ -3,14 +3,14 @@ using Comrade
 using PyCall
 using Plots
 
-using GalacticOptim, Optim
 using Zygote
-using Parameters
 using Distributions, DistributionsAD
-using AdvancedHMC
-using AbstractDifferentiation
-using ReverseDiff
+using ComradeAHMC
+using ComradeGalactic
+using GalacticBBO
 using ForwardDiff
+using GalacticOptimJL
+using DistributionsAD
 
 load_ehtim()
 @pyimport eht_dmc as ed
@@ -19,91 +19,96 @@ load_ehtim()
 obs = ehtim.obsdata.load_uvfits(joinpath(@__DIR__, "SR1_M87_2017_096_hi_hops_netcal_StokesI.uvfits"))
 obs.add_scans()
 
-obsavg = obs.flag_uvdist(uv_min=0.1e9).avg_coherent(0.0, scan_avg=true).add_fractional_noise(0.01)
+obsavg = obs.flag_uvdist(uv_min=0.1e9).avg_coherent(0.0, scan_avg=true).add_fractional_noise(0.02)
 
 obslist = obsavg.split_obs()
 
-dvis = extract_vis(ehtim.obsdata.merge_obs(obslist[1:20]))
+dvis = extract_vis(ehtim.obsdata.merge_obs(obslist))
 st = scantable(dvis)
 
-gcache = Comrade.GainCache(st)
+gcache = GainCache(st)
 
-function model(θ, cache, gcache, fovx, fovy, ny, nx)
-    @unpack c, f, gamp, gphase = θ
-    g = gamp.*cis.(gphase)
-    cmat = reshape(c, ny, nx)
-    img = IntensityMap(cmat, fovx, fovy, BSplinePulse{3}())
-    m = f*modelimage(img, cache)
-    Comrade.GainModel(gcache, g, m)
+struct Model{G,C}
+    gcache::G
+    cache::C
+    fovx::Float64
+    fovy::Float64
+    nx::Int
+    ny::Int
 end
+
+function (model::Model)(θ)
+    (;c, gamp, gphase) = θ
+    g = exp.(gamp).*cis.(gphase)
+    cmat = reshape(c, model.ny, model.nx)
+    img = IntensityMap(cmat, model.fovx, model.fovy, BSplinePulse{3}())
+
+    m = modelimage(img, model.cache)
+    return GainModel(model.gcache, g, m)
+end
+
 # define the priors
-distamp = (AA = LogNormal(0.0, 0.01),
-         AP = LogNormal(0.0, 0.1),
-         LM = LogNormal(0.0, 0.5),
-         AZ = LogNormal(0.0, 0.1),
-         JC = LogNormal(0.0, 0.1),
-         PV = LogNormal(0.0, 0.1),
-         SM = LogNormal(0.0, 0.1)
+distamp = (AA = Normal(0.0, 1e-3),
+         AP = Normal(0.0, 0.1),
+         LM = Normal(0.0, 0.5),
+         AZ = Normal(0.0, 0.1),
+         JC = Normal(0.0, 0.1),
+         PV = Normal(0.0, 0.1),
+         SM = Normal(0.0, 0.1)
          )
-distphase = (AA = VonMises(0.0, 1e4),
-             AP = VonMises(0.0, 1/π^2),
-             LM = VonMises(0.0, 1/π^2),
-             AZ = VonMises(0.0, 1/π^2),
-             JC = VonMises(0.0, 1/π^2),
-             PV = VonMises(0.0, 1/π^2),
-             SM = VonMises(0.0, 1/π^2)
+distphase = (AA = Normal(0.0, 0.001),
+             AP = Normal(0.0, 1π),
+             LM = Normal(0.0, 1π),
[…]
