+++ examples/gaintest.jl
@@ -3,13 +3,7 @@ using Comrade
 using PyCall
 using Plots
 
-using GalacticOptim, Optim
-using Zygote
-using Parameters
-using Distributions
-using SpeedMapping
-using AdvancedHMC
-using Dynesty
+using ComradeAHMC
 
 load_ehtim()
 @pyimport eht_dmc as ed
@@ -34,43 +28,43 @@ gainp = Dict("AA"=> 0.01,
              "PV" => 0.1,
              "SM" => 0.1,
              "SR" => 0.1)
-obsim = img.observe_same(ehtim.obsdata.merge_obs(obslist[10:15]), ttype="fast",  gainp=gainp, gain_offset=gainoff, ampcal=false, phasecal=false, stabilize_scan_phase=true, stabilize_scan_amp=true)
+obsim = img.observe_same(ehtim.obsdata.merge_obs(obslist[1:15]), ttype="fast",  gainp=gainp, gain_offset=gainoff, ampcal=false, phasecal=false, stabilize_scan_phase=true, stabilize_scan_amp=true)
 # now get cal table
 ctable = ehtim.calibrating.self_cal.self_cal(obsim, img, caltable=true)
 ehtim.caltable.save_caltable(ctable, obsim, datadir=joinpath(@__DIR__ ,"CaltableTest"))
 
 
 dvis = extract_vis(obsim)
-dlcamp = extract_lcamp(obsim; count="min")
-dcphase = extract_cphase(obsim, count="min-cut0bl")
+dlcamp = extract_lcamp(obsim)
+dcphase = extract_cphase(obsim)
 damp = extract_amp(obsim)
 st = scantable(dvis)
 
 gcache = Comrade.GainCache(st)
 
-struct Model{G}
+struct Test{G}
     gcache::G
 end
 
-function Model(st::Comrade.ScanTable)
+function Test(st::Comrade.ScanTable)
     gcache = Comrade.GainCache(st)
-    return Model{typeof(gcache)}(gcache)
+    return Test{typeof(gcache)}(gcache)
 end
 
-function (mod::Model)(θ)
-    @unpack f, σ, τ, ξ,  gamp, gphase = θ
-    g = gamp.*cis.(gphase)
-    m = f*rotated(stretched(Gaussian(), σ*τ, σ), ξ)
+function (mod::Test)(θ)
+    (;f, σ, τ, ξ,  gamp, gphase) = θ
+    g = exp.(gamp).*cis.(gphase)
+    m = f*Comrade.rotate(stretch(Gaussian(), σ*τ, σ), ξ)
     Comrade.GainModel(mod.gcache, g, m)
 end
 
 # define the priors
-distamp = (AA = LogNormal(0.0, 0.1),
-           AP = LogNormal(0.0, 0.1),
-           LM = LogNormal(0.0, 0.2),
-           AZ = LogNormal(0.0, 0.1),
-           JC = LogNormal(0.0, 0.1),
-           PV = LogNormal(0.0, 0.1),
+distamp = (AA = Normal(0.0, 0.1),
+           AP = Normal(0.0, 0.1),
+           LM = Normal(0.0, 0.2),
+           AZ = Normal(0.0, 0.1),
+           JC = Normal(0.0, 0.1),
+           PV = Normal(0.0, 0.1),
            #SM = LogNormal(0.0, 0.1)
          )
 distphase = (AA = Normal(0.0, 1e-4),
@@ -91,7 +85,7 @@ prior = (
         )
 # Now form the posterior
 
-mms  = Model(st)
+mms  = Test(st)
 
 cllklhd = RadioLikelihood(dlcamp, dcphase)
[…]

[…]
