### 2022-06-15 5898d2c
add laplace method
