+++ lib/ComradeGalactic/src/ComradeGalactic.jl
@@ -2,9 +2,12 @@ module ComradeGalactic
 
 using Comrade
 using Reexport
+using Distributions
+using ForwardDiff
 
 @reexport using GalacticOptim
 
+export laplace
 
 
 function GalacticOptim.OptimizationFunction(post::Comrade.Posterior, args...; kwargs...)
@@ -15,6 +18,25 @@ function GalacticOptim.OptimizationFunction(post::Comrade.TransformedPosterior, 
     ℓ(x,p) = -logdensityof(post, x)
     return OptimizationFunction(ℓ, args...; kwargs...)
 end
+
+"""
+    `laplace(prob, opt, args...; kwargs...)`
+Compute the Laplace or Quadratic approximation to the prob or posterior.
+The `args` and `kwargs` are passed the the GalacticOptim.solve function.
+
+Note the quadratic approximation is in the space of the transformed posterior
+not the usual parameter space. This is better for constrained problems where
+we may run up against a boundary.
+"""
+function laplace(prob::OptimizationProblem, opt, args...; kwargs...)
+    sol = solve(prov, opt, args...; kwargs...)
+    f = prob.f.f
+    J = -ForwardDiff.hessian(f, sol)
+    h = hess*sol
+    return MvNormalCanon(h, Symmetric(J))
+end
+
+
 
 # """
 #     solve(prob, opt, args...; transform=nothing, kwargs...)
