### 2022-06-15 68c1c1e
CompatHelper: bump compat for MeasureBase to 0.10, (keep existing compat)
