+++ Project.toml
@@ -69,7 +69,7 @@ FileIO = "1"
 ForwardDiff = "0.9, 0.10"
 HypercubeTransform = "^0.2.7"
 KeywordCalls = "0.2"
-MeasureBase = "0.7, 0.8, 0.9"
+MeasureBase = "0.7, 0.8, 0.9, 0.10"
 MeasureTheory = "0.15, 0.16"
 NFFT = "0.10, 0.11, 0.12"
 NamedTupleTools = "0.13, 0.14"
