+++ examples/gain_m87.jl
@@ -10,7 +10,6 @@ using ComradeGalactic
 using GalacticBBO
 using ForwardDiff
 using GalacticOptimJL
-using DistributionsAD
 
 load_ehtim()
 @pyimport eht_dmc as ed
@@ -22,9 +21,10 @@ obs.add_scans()
 obsavg = obs.flag_uvdist(uv_min=0.1e9).avg_coherent(0.0, scan_avg=true).add_fractional_noise(0.02)
 
 obslist = obsavg.split_obs()
-
-dvis = extract_vis(ehtim.obsdata.merge_obs(obslist))
-st = scantable(dvis)
+obss = ehtim.obsdata.merge_obs(obslist)
+damp = extract_amp(obss)
+dcp = extract_cphase(obss)
+st = scantable(damp)
 
 gcache = GainCache(st)
 
@@ -38,8 +38,8 @@ struct Model{G,C}
 end
 
 function (model::Model)(θ)
-    (;c, gamp, gphase) = θ
-    g = exp.(gamp).*cis.(gphase)
+    (;c, gamp) = θ
+    g = exp.(gamp)
     cmat = reshape(c, model.ny, model.nx)
     img = IntensityMap(cmat, model.fovx, model.fovy, BSplinePulse{3}())
 
@@ -48,15 +48,15 @@ function (model::Model)(θ)
 end
 
 # define the priors
-distamp = (AA = Normal(0.0, 1e-3),
+distamp = (AA = Normal(0.0, 0.1),
          AP = Normal(0.0, 0.1),
-         LM = Normal(0.0, 0.5),
+         LM = Normal(0.0, 1.0),
          AZ = Normal(0.0, 0.1),
          JC = Normal(0.0, 0.1),
          PV = Normal(0.0, 0.1),
          SM = Normal(0.0, 0.1)
          )
-distphase = (AA = Normal(0.0, 0.001),
+distphase = (AA = Normal(0.0, 0.0001),
              AP = Normal(0.0, 1π),
              LM = Normal(0.0, 1π),
              AZ = Normal(0.0, 1π),
@@ -65,50 +65,50 @@ distphase = (AA = Normal(0.0, 0.001),
              SM = Normal(0.0, 1π)
              )
 # Now form the posterior
-npix = 8
+npix = 12
 prior = (
-          c = MvLogNormal(fill(-log(10.0*npix*npix), npix*npix), fill(1.0, npix*npix)),
+          c = MvLogNormal(fill(-log(4*npix*npix), npix*npix), fill(2.0, npix*npix)),
           #c = DistributionsAD.TuringDirichlet(npix*npix, 1.0),
           #f = Uniform(0.0, 1.0),
           gamp = Comrade.GainPrior(distamp, st),
-          gphase = Comrade.GainPrior(distphase, st)
         )
 
 # Now form the posterior
 
-fovxy = μas2rad(65.0)
-gcache = GainCache(st)
-cache = create_cache(Comrade.DFTAlg(dvis), IntensityMap(zeros(npix,npix), fovxy, fovxy, BSplinePulse{3}()))
+fovxy = μas2rad(70.0)
+cache = create_cache(Comrade.DFTAlg(damp), IntensityMap(zeros(npix,npix), fovxy, fovxy, BSplinePulse{3}()))
 mms = Model(gcache, cache, fovxy, fovxy, npix, npix)
 
 #mms = NoGModel(cache, fovxy, fovxy, npix, npix)
-lklhd = RadioLikelihood(dvis)
+lklhd = RadioLikelihood(damp, dcp)
 post = Posterior(lklhd, prior, mms)
 tpost = asflat(post)
+lp = logdensityof(tpost)
+
[…]
