+++ lib/ComradeGalactic/Project.toml
@@ -12,6 +12,8 @@ Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 
 [compat]
 Comrade = "0.3,0.4"
+Distributions = "0.24,0.25"
+ForwardDiff = "0.9, 0.10"
 GalacticOptim = "3"
 Reexport = "1"
 julia = "1.6"
