+++ lib/ComradeBase/test/runtests.jl
@@ -1,0 +1,6 @@
+using ComradeBase
+using Test
+
+@testset "ComradeBase.jl" begin
+    # Write your tests here.
+end
