+++ lib/ComradeBase/README.md
@@ -1,0 +1,3 @@
+# ComradeBase
+
+Base interface package for `Comrade` this is a lighter dependency the just defines some base methods.

[…]
