### 2022-06-15 e3bac0f
Add more gain tests
