+++ test/Core/gains.jl
@@ -15,11 +15,13 @@ using Distributions
          σ1 = 1.2e-10,
          τ1 = 0.25,
          ξ1 = -0.42,
-         f2 = 0.5, σ2 = 8.0e-12,
+         f2 = 0.5,
+         σ2 = 8.0e-12,
          τ2 = 0.4,
          ξ2 = 0.0,
          x = 4.0e-11,
-         y = 1.5e-10)
+         y = 1.5e-10
+        )
 
     m = test_model(θ)
 
