### 2022-06-16 7f1aac9
fix data.jl docs example
