+++ docs/src/examples/data.jl
@@ -4,7 +4,7 @@
 # particular uvfits format that is similar to the optical interferometry oifits format.
 # As a result, we reuse the excellent `eht-imaging` package to load data into `Comrade`.
 
-# Once the data is loaded we then load the data into specific data formats that `Comrade`
+# Once the data is loaded we then convert the data into the tabular format `Comrade`
 # expects. Note that in the future this may change to a Julia package as the Julia radio
 # astronomy group grows.
 
@@ -14,7 +14,7 @@ using Comrade
 using Plots
 
 # Now we load ehtim. This assumes you have a working installation of `eht-imaging`.
-# To install eht-imaging see the [ehtim](https://github.com/achael/ehtim) github repo.
+# To install eht-imaging see the [ehtim](https://github.com/achael/eht-imaging) github repo.
 
 load_ehtim()
 
@@ -30,8 +30,8 @@ obs = obs.avg_coherent(0.0, scan_avg=true)
 # We can now extract data products that `Comrade` can use
 vis = extract_vis(obs) #complex visibilites
 amp = extract_amp(obs) # visibility amplitudes
-cphase = extract_cphase(obs; count="min") # extract minimal set of closure phases
-lcamp = extract_lcamp(obs; count="min") # extract minimal set of log-closure amplitudes
+cphase = extract_cphase(obs) # extract minimal set of closure phases
+lcamp = extract_lcamp(obs) # extract minimal set of log-closure amplitudes
 
 # We can also recover the array used in the observation using
 ac = arrayconfig(vis)
@@ -39,7 +39,7 @@ plot(ac) # Plot the baseline coverage
 
 # To plot the data we just call
 
-l = @layout [ a b; c d]
+l = @layout [a b; c d]
 pv = plot(vis)
 pa = plot(amp)
 pcp = plot(cphase)
