### 2022-06-16 be15018
fix up black_hole_image.jl
