### 2022-06-16 1d6918a
Fix nonanalytic.jl with new ExtendedRing model.
