### 2022-06-16 34bb82b
fix typos in docs
