+++ lib/ComradeGalactic/test/runtests.jl
@@ -1,4 +1,4 @@
-using Comrade, ComradeGalactic, GalacticBBO, Distributions
+using Comrade, ComradeGalactic, GalacticOptimJL, Distributions
 using Test
 
 load_ehtim()
@@ -12,9 +12,19 @@ include(joinpath(@__DIR__, "../../../test/test_util.jl"))
 
     tpost = ascube(post)
     f = OptimizationFunction(tpost, GalacticOptim.AutoForwardDiff{4}())
-    prob = GalacticOptim.OptimizationProblem(f, prior_sample(tpost), nothing, lb=fill(0.001, 10), ub=fill(0.999,10))
-    sol = solve(prob, BBO_adaptive_de_rand_1_bin(); maxiters=100_000)
-#
+    x0 = [ 0.1,
+           0.4,
+           0.5,
+           0.1,
+           0.3,
+           0.1,
+           0.4,
+           0.3,
+           0.8,
+           0.8]
+    prob = GalacticOptim.OptimizationProblem(f, x0, nothing, lb=fill(0.001, 10), ub=fill(0.999,10))
+    sol = solve(prob, LBFGS(); maxiters=100_000)
+
     xopt = transform(tpost, sol)
     @test isapprox(xopt.f1/xopt.f2, 2.0, atol=1e-3)
     @test isapprox(xopt.σ1*2*sqrt(2*log(2)), μas2rad(40.0), rtol=1e-3)
