### 2022-06-17 8b8a186
speed up Composite model and fix a dynamic dispatch
