### 2022-06-20 bd6a01f
CompatHelper: bump compat for MeasureBase to 0.11, (keep existing compat)
