### 2022-06-21 34d1025
CompatHelper: bump compat for MeasureBase to 0.12, (keep existing compat)
