### 2022-06-21 77ca455
Update README.md
Removed roadmap since this should mostly be covered in #18
