+++ README.md
@@ -19,11 +19,11 @@ add Comrade
 ```
 To perform inference i.e. parameter estimation you should look at the `lib` packages. 
 
-# Roadmap
+<!-- # Roadmap
 
 1. Settle on model interface
 2. Design a better data maniuplation interface and native uvfits reading (or spin up the EHTIM.jl package)
 2. ~Full gain~ and polarization implemented, including DTERM modeling ala DMC, and approximate methods such as pseudo-marginal, particle MCMC methods, or laplace approximations
 3. Write model adjoints when needed for additional speed (and define a may to define AD backends like Zygote, Enzyme, Diffractor, etc.)
 4. GPU support when its a good idea (maybe for gain solves?)
-
+ -->
