### 2022-06-21 b761db7
start fixing docstrings
