+++ README.md
@@ -3,6 +3,7 @@
 [![Dev](https://img.shields.io/badge/docs-dev-blue.svg)](https://ptiede.github.io/Comrade.jl/dev/)
 [![Build Status](https://github.com/ptiede/Comrade.jl/workflows/CI/badge.svg)](https://github.com/ptiede/Comrade.jl/actions)
 [![Coverage](https://codecov.io/gh/ptiede/Comrade.jl/branch/main/graph/badge.svg)](https://codecov.io/gh/ptiede/Comrade.jl)
+[![ColPrac: Contributor's Guide on Collaborative Practices for Community Packages](https://img.shields.io/badge/ColPrac-Contributor's%20Guide-blueviolet)](https://github.com/SciML/ColPrac)
 
 Composable Modeling of Radio Emission
 
