### 2022-06-21 d2b325a
fix up docstrings in Bayes
