+++ src/bayes/bayes.jl
@@ -7,25 +7,33 @@ using HypercubeTransform
 using TransformVariables
 using ValueShapes: NamedTupleDist
 
-"""
-    Posterior(lklhd, prior, model)
-Posterior density that follows obeys the [DensityInferface](https://github.com/JuliaMath/DensityInferface.jl)
-
-The expected arguments are:
-
-- lklhd: Which should be an intance of RadioLikelihood with whatever data products you want to fit
-- prior: This should be a `NamedTuple` with the priors for each model ParameterHandling
-- model: Function that takes a `NamedTuple` of parameters and constructs the `Comrade` model.
-
-To evaluate the logdensity of the posterior use can either use `DensityInterface.logdensityof`
-or `logdensity`.
-"""
 struct Posterior{L,P,F}
     lklhd::L
     prior::P
     model::F
 end
 
+"""
+    Posterior(lklhd, prior, model)
+Creates a Posterior density that follows obeys [DensityInferface](https://github.com/JuliaMath/DensityInferface.jl).
+The `lklhd` object is expected to be a `MeasureBase.Likelihood` object. For instance, these can be
+created using [`RadioLikelihood](@ref RadioLikelihood). `prior` is expected to be a `NamedTuple`
+of distributions that reflect the priors on the parameters you are considering. `model` is a function
+that takes in a `NamedTuple` of parameters and returns a `Comrade` `<:AbstractModel`.
+
+# Notes
+Since this function obeys [`DensityInferface`](@ref DensityInterface) you can evaluate it with
+```julia
+ℓ = logdensityof(post)
+ℓ(x)
+```
+or using the 2-argument version directly
+```julia
+logdensityof(post, x)
+```
+
+To generate random draws from the prior see the [`prior_sample`](@ref prior_sample) function.
+"""
 function Posterior(lklhd, prior::NamedTuple, model)
     return Posterior(lklhd, NamedTupleDist(prior), model)
 end
@@ -39,31 +47,30 @@ function DensityInterface.logdensityof(post::Posterior, x)
 end
 
 """
-    $(SIGNATURES)
-Samples the prior distribution from the posterior `nsamples` times.
+    prior_sample(post::Posterior, args...)
 
-Returns a Vector{NamedTuple} that can be used to initialize optimization and sample algorithms
+Samples the prior distribution from the posterior. The `args...` are forwarded to the
+[`Base.rand`](@ref Base.rand) method.
 """
 function prior_sample(post::Posterior, args...)
     return rand(post.prior, args...)
 end
 
+"""
+    prior_sample(post::Posterior)
+
+Returns a single sample from the prior distribution.
+"""
 function prior_sample(post::Posterior)
     return rand(post.prior)
 end
 
 """
-    `$(TYPEDEF)`
-A transformed version of the posterior `lpost`. This is an internal type
-that an end user shouldn't have to directly construct.
-
-To construct a transformed posterior see the `asflat`, `ascube`, and `flatten`
-functions.
-
-When evaluating the `logdensity` of the `TransformedPosterior`, `logdenisty`
-will include any jacobian terms automatically, and expects the argument to be the
-vector of transformed quantities. This is usually easier to incorporate with
-samplers, and optimizers, which expect homogenous vectors as arguments.
+    $(TYPEDEF)
+A transformed version of a `Posterior` object.
+This is an internal type that an end user shouldn't have to directly construct.
+To construct a transformed posterior see the [`asflat`](@ref asflat), [`ascube`](@ref ascube),
+and [`flatten`](@ref Comrade.flatten) docstrings.
 """
 struct TransformedPosterior{P<:Posterior,T}
     lpost::P
@@ -81,38 +88,57 @@ function prior_sample(tpost::TransformedPosterior)
 end
 
 
[…]
