+++ src/models/modelimage/modelimage.jl
@@ -2,11 +2,12 @@ export modelimage
 
 """
     $(TYPEDEF)
-Container for non-analytic model that contains a image cache which will hold the image,
-a Fourier transform cache C, which usually an instance of a <: FourierCache which usually
-holds a interpolator that allows you to compute visibilities.
 
-# Notes
+Container for non-analytic model that contains a image cache which will hold the image,
+a Fourier transform cache, which usually an instance of a <: FourierCache.
+
+# Note
+
 This is an internal implementation detail that shouldn't usually be called directly.
 Instead the user should use the exported function `modelimage`, for example
 
@@ -21,6 +22,10 @@ mimg = modelimage(m) # you can also optionally pass the number of pixels nx and 
 # Or you can create an IntensityMap
 img = intensitymap(m, 100.0, 100.0, 512, 512)
 mimg = modelimage(m, img)
+
+# Or precompute a cache
+cache = create_cache(FFTAlg(), img)
+mimg = modelimage(m, cache)
 ```
 """
 struct ModelImage{M,I,C} <: AbstractModelImage{M}
@@ -64,12 +69,13 @@ include(joinpath(@__DIR__, "cache.jl"))
 """
     $(SIGNATURES)
 Construct a `ModelImage` from a `model`, `image` and the optionally
-specified visibility algorithm `alg`
+specified visibility algorithm `alg` and `executor` which uses Folds.jl
+parallelism formalism.
 
 # Notes
-For analytic models this is a no-op and just return the model.
-For non-analytic models this wraps the model in a object with an image
-and precomputes the fourier transform using `alg`.
+For analytic models this is a no-op and returns the model.
+For non-analytic models this creates a `ModelImage` object which uses `alg` to compute
+the non-analytic Fourier transform.
 """
 @inline function modelimage(model::M, image::ComradeBase.AbstractIntensityMap, alg::FourierTransform=FFTAlg(), executor=SequentialEx()) where {M}
     return modelimage(visanalytic(M), model, image, alg, executor)
@@ -92,13 +98,29 @@ end
     _modelimage(model, image, alg, executor)
 end
 
+"""
+    $(SIGNATURES)
+
+Construct a `ModelImage` from the `model` and using a precompute Fourier transform `cache`.
+You can optionally specify the executor which will compute the internal image buffer using
+the `executor`.
+
+# Example
+
+```julia
+m = ExtendedRing(10.0)
+cache = create_cache(DFTAlg(), IntensityMap(zeros(128, 128), 50.0, 50.0)) # used threads to make the image
+mimg = modelimage(m, cache, ThreadedEx())
+```
+
+# Notes
+For analytic models this is a no-op and returns the model.
+
+"""
 @inline function modelimage(model::M, cache::AbstractCache, executor=SequentialEx()) where {M}
     return modelimage(visanalytic(M), model, cache, executor)
 end
 
-@inline function modelimage(::IsAnalytic, model, cache::AbstractCache, args...)
-    return model
-end
 
 @inline function modelimage(::NotAnalytic, model, cache::AbstractCache, executor=SequentialEx())
     img = cache.img
@@ -107,11 +129,32 @@ end
     return ModelImage(model, img, newcache)
 end
 
+"""
+    $(SIGNATURES)
+
+Create a model image directly using an image, i.e. treating it as the model. You
+can optionally specify the Fourier transform algorithm using `alg`
+
+# Notes
+For analytic models this is a no-op and returns the model.
+
+"""
 @inline function modelimage(img::IntensityMap, alg=NFFTAlg())
     cache = create_cache(alg, img)
     return ModelImage(img, img, cache)
 end
 
+"""
+    $(SIGNATURES)
+
+Create a model image directly using an image, i.e. treating it as the model. Additionally
[…]
