+++ Project.toml
@@ -85,7 +85,7 @@ StatsBase = "0.31,0.32,0.33"
 StructArrays = "0.3,0.4,0.5,0.6"
 Tables = "1"
 TransformVariables = "0.5,0.6"
-TupleVectors = "0.1"
+TupleVectors = "0.1, 0.2"
 ValueShapes = "0.10"
 julia = "^1.6"
 
