+++ lib/ComradeDynesty/Project.toml
@@ -15,7 +15,7 @@ AbstractMCMC = "3, 4"
 Comrade = "0.3,0.4"
 Dynesty = "0.1"
 Reexport = "1"
-TupleVectors = "0.1"
+TupleVectors = "0.1, 0.2"
 julia = "1.6"
 
 [extras]
