+++ lib/ComradeOptimization/Project.toml
@@ -1,7 +1,7 @@
-name = "ComradeGalactic"
+name = "ComradeOptimization"
 uuid = "47701775-187b-49a4-b924-794639d71cdc"
 authors = ["Paul Tiede <ptiede91@gmail.com>"]
-version = "0.1.2"
+version = "0.1.0"
 
 [deps]
 Comrade = "99d987ce-9a1e-4df8-bc0b-1ea019aa547b"
@@ -17,7 +17,7 @@ Distributions = "0.24,0.25"
 ForwardDiff = "0.9, 0.10"
 GalacticOptim = "3"
 Reexport = "1"
-julia = "1.6"
+julia = "1.7"
 
 [extras]
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
@@ -26,4 +26,4 @@ GalacticOptimJL = "9d3c5eb1-403b-401b-8c0f-c11105342e6b"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Test", "GalacticBBO", "GalacticOptimJL", "Distributions"]
+test = ["Test", "OptimizationBBO", "OptimizationOptimJL", "Distributions"]
