+++ lib/ComradeOptimization/Project.toml
@@ -1,5 +1,5 @@
 name = "ComradeOptimization"
-uuid = "47701775-187b-49a4-b924-794639d71cdc"
+uuid = "26988f03-70df-4135-959e-06268cc6415a"
 authors = ["Paul Tiede <ptiede91@gmail.com>"]
 version = "0.1.0"
 
