### 2022-07-12 a04b45c
fix problems with the project
