+++ lib/ComradeOptimization/src/ComradeOptimization.jl
@@ -1,36 +1,37 @@
-module Optimization
+module ComradeOptimization
 
 using Comrade
 using Reexport
 using Distributions
 using ForwardDiff
 using LinearAlgebra
+import SciMLBase
 
-@reexport using Optimziation
+@reexport using Optimization
 
 export laplace
 
 
-function GalacticOptim.OptimizationFunction(post::Comrade.Posterior, args...; kwargs...)
+function SciMLBase.OptimizationFunction(post::Comrade.Posterior, args...; kwargs...)
     throw("Transform the posterior first using `asflat` or `ascube`")
 end
 
-function GalacticOptim.OptimizationFunction(post::Comrade.TransformedPosterior, args...; kwargs...)
+function SciMLBase.OptimizationFunction(post::Comrade.TransformedPosterior, args...; kwargs...)
     ℓ(x,p) = -logdensityof(post, x)
-    return OptimizationFunction(ℓ, args...; kwargs...)
+    return SciMLBase.OptimizationFunction(ℓ, args...; kwargs...)
 end
 
 """
     laplace(prob, opt, args...; kwargs...)
 
 Compute the Laplace or Quadratic approximation to the prob or posterior.
-The `args` and `kwargs` are passed the the GalacticOptim.solve function.
+The `args` and `kwargs` are passed the the SciMLBase.solve function.
 
 Note the quadratic approximation is in the space of the transformed posterior
 not the usual parameter space. This is better for constrained problems where
 we may run up against a boundary.
 """
-function laplace(prob::OptimizationProblem, opt, args...; kwargs...)
+function laplace(prob::SciMLBase.OptimizationProblem, opt, args...; kwargs...)
     sol = solve(prob, opt, args...; kwargs...)
     f = Base.Fix2(prob.f, nothing)
     J = ForwardDiff.hessian(f, sol)
@@ -50,7 +51,7 @@ end
 # a tuple with the galactic optim solution in the first element and the optimum
 # location in model space in the second argument
 # """
-# function GalacticOptim.solve(prob::OptimizationProblem, opt, transform::Union{Nothing, HypercubeTransform.AbstractHypercubeTransform, HypercubeTransform.TransformVariables.AbstractTransform, HypercubeTransform.NamedFlatTransform}, args...; kwargs...)
+# function SciMLBase.solve(prob::OptimizationProblem, opt, transform::Union{Nothing, HypercubeTransform.AbstractHypercubeTransform, HypercubeTransform.TransformVariables.AbstractTransform, HypercubeTransform.NamedFlatTransform}, args...; kwargs...)
 #     sol = solve(prob, opt, args...; kwargs...)
 #     if isnothing(transform)
 #         return sol
