+++ lib/ComradeOptimization/Project.toml
@@ -7,21 +7,22 @@ version = "0.1.0"
 Comrade = "99d987ce-9a1e-4df8-bc0b-1ea019aa547b"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
-Optimization = "a75be94c-b780-496d-a8a9-0878b188d577"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
+Optimization = "7f7a1694-90dd-40f0-9382-eb1efda571ba"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
+SciMLBase = "0bca4576-84f4-4d90-8ffe-ffa030f20462"
 
 [compat]
 Comrade = "0.3,0.4"
 Distributions = "0.24,0.25"
 ForwardDiff = "0.9, 0.10"
-Optimization = "3"
 Reexport = "1"
+SciMLBase = "1"
 julia = "1.7"
 
 [extras]
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
-OptimizationBBO = "80c49c3a-6557-47d9-8f5b-13d0a2920315"
+OptimizationBBO = "3e6eede4-6085-4f62-9a71-46d9bc1eb92b"
 OptimizationOptimJL = "9d3c5eb1-403b-401b-8c0f-c11105342e6b"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
