+++ lib/ComradeOptimization/Project.toml
@@ -16,6 +16,7 @@ SciMLBase = "0bca4576-84f4-4d90-8ffe-ffa030f20462"
 Comrade = "0.3,0.4"
 Distributions = "0.24,0.25"
 ForwardDiff = "0.9, 0.10"
+Optimization = "3"
 Reexport = "1"
 SciMLBase = "1"
 julia = "1.7"
