+++ src/models/models.jl
@@ -400,3 +400,4 @@ include(joinpath(@__DIR__, "combinators.jl"))
 include(joinpath(@__DIR__, "geometric_models.jl"))
 include(joinpath(@__DIR__, "polarized.jl"))
 include(joinpath(@__DIR__, "test.jl"))
+include(joinpath(@__DIR__, "misc.jl"))
