+++ Project.toml
@@ -21,6 +21,7 @@ FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
 FITSIO = "525bcba6-941b-5504-bd06-fd0dc1a4d2eb"
 FLoops = "cc61a311-1640-44b5-9fba-1b764f453329"
 FileIO = "5789e2e9-d7fb-5bc7-8068-2c6fae9b9549"
+FillArrays = "1a297f60-69ca-5386-bcde-b61e274b549b"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
 HypercubeTransform = "9ec9aee3-0fd3-44c2-8e61-a50acc66f3c8"
 KeywordCalls = "4d827475-d3e4-43d6-abe3-9688362ede9f"
@@ -66,6 +67,7 @@ FFTW = "^1"
 FITSIO = "0.16"
 FLoops = "0.2"
 FileIO = "1"
+FillArrays = "0.13"
 ForwardDiff = "0.9, 0.10"
 HypercubeTransform = "^0.2.7"
 KeywordCalls = "0.2"
