+++ .github/workflows/CompatHelper.yml
@@ -13,4 +13,4 @@ jobs:
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           COMPATHELPER_PRIV: ${{ secrets.DOCUMENTER_KEY }}
-        run: julia -e 'using CompatHelper; CompatHelper.main(subdirs = ["", "docs", "lib/ComradeAdaptMCMC", "lib/ComradeAHMC", "lib/ComradeDynesty", "lib/ComradeGalactic", "lib/ComradeNested"])'
+        run: julia -e 'using CompatHelper; CompatHelper.main(subdirs = ["", "docs", "lib/ComradeAdaptMCMC", "lib/ComradeAHMC", "lib/ComradeDynesty", "lib/ComradeOptimization", "lib/ComradeNested"])'
