### 2022-07-12 4adc98d
Fix stack overflow in model
