+++ docs/src/benchmarks.md
@@ -1,0 +1,155 @@
+# Benchmarks
+
+`Comrade` was partially designed with performance in mind. Solving imaging inverse problem is traditionally very computational expensive, especially when using Bayesian inference. However, `Comrade` also was designed to allow a user to easily modify the image model and refit the data with few changes to any script or notebook. 
+
+To benchmark `Comrade` we will compare it to two of the most common modeling or imaging packages within the EHT:
+
+- [eht-imaging](https://github.com/achael/eht-imaging/)
+- [Themis](https://iopscience.iop.org/article/10.3847/1538-4357/ab91a4)
+
+`eht-imaging`[^1] or `ehtim` is a Python package that is widely used within the EHT for its imaging and modeling interfaces. It is easy to use and is reasonably fast due, however, this comes with the caveat that the user must specify each model separately, and provide handwritten gradient definitions. Additionally, `ehtim`'s modeling interface is rather new, since it is mostly used to produce images using a method called *regularized maximum likelihood*, which is similar to `Comrade`'s approach except it only characterizes the images using the maximum likelihood estimate.
+
+Themis is a C++ package that is focused on providing Bayesian estimates of the image structure. In fact, `Comrade` took some design queue's from `Themis`. However, `Themis` is quite challenging to use and requires a high level of knowledge from its users, requiring them to understand makefile, C++, and the MPI standard. Additionally, Themis was designed to solely work with distributed computing systems. Unfortunately, as of the writing of this document, Themis is still closed source software and it only available to members of the EHT.
+
+## Benchmarking Problem
+
+For our benchmarking problem we analyze a problem very similar to the problem explained in [Making an Image of a Black Hole](@ref). Namely we will consider fitting the 2017 M87 April 6 data using an m-ring and a single Gaussian component. To see the code we used for `Comrade` and `eht-imaging` please see the end of this page.
+
+
+# Results
+
+All tests were run using the following Julia system
+
+Julia Version 1.7.3
+Python Version 3.10.5
+Commit 742b9abb4d (2022-05-06 12:58 UTC)
+Platform Info:
+  OS: Linux (x86_64-pc-linux-gnu)
+  CPU: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
+  WORD_SIZE: 64
+  LIBM: libopenlibm
+  LLVM: libLLVM-12.0.1 (ORCJIT, tigerlake)
+
+Our benchmark results are the following:
+
+| | Comrade (micro sec) | eht-imaging (micro sec) | Themis (micro sec)|
+|---|---|---|---|
+| posterior eval (min) | 31  | 445  | 55  |
+| posterior eval (mean) | 36  | 476  | 60  |
+| grad posterior eval (min) |  105 (ForwardDiff) | 1898  | 1809  |
+| grad posterior eval (mean) |  119 (ForwardDiff) | 1971 |  1866  |
+
[…]
