+++ docs/Project.toml
@@ -1,30 +1,30 @@
 [deps]
+Comrade = "99d987ce-9a1e-4df8-bc0b-1ea019aa547b"
 ComradeAHMC = "a4336a5c-78bc-4363-8a90-ce3fa9d3abe4"
 ComradeBase = "6d8c423b-a35f-4ef1-850c-862fe21f82c4"
-ComradeGalactic = "47701775-187b-49a4-b924-794639d71cdc"
+ComradeOptimization = "26988f03-70df-4135-959e-06268cc6415a"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
-GalacticBBO = "80c49c3a-6557-47d9-8f5b-13d0a2920315"
-GalacticOptim = "a75be94c-b780-496d-a8a9-0878b188d577"
-GalacticOptimJL = "9d3c5eb1-403b-401b-8c0f-c11105342e6b"
 Glob = "c27321d9-0574-5035-807b-f59d2c89b15c"
 HypercubeTransform = "9ec9aee3-0fd3-44c2-8e61-a50acc66f3c8"
 Literate = "98b081ad-f1c9-55d3-8b20-4c87d4299306"
+Optimization = "7f7a1694-90dd-40f0-9382-eb1efda571ba"
+OptimizationBBO = "3e6eede4-6085-4f62-9a71-46d9bc1eb92b"
+OptimizationOptimJL = "36348300-93cb-4f02-beb5-3c3902f8871e"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 
 [compat]
 ComradeAHMC = "0.1"
 ComradeBase = "0.3"
-ComradeGalactic = "0.1"
+ComradeOptimization = "0.1"
 Distributions = "0.25"
 Documenter = "0.27"
 ForwardDiff = "0.10"
-GalacticBBO = "0.1"
-GalacticOptim = "3"
-GalacticOptimJL = "0.1"
 Glob = "1"
 HypercubeTransform = "^0.2.7"
 Literate = "2"
+Optimization = "3"
+OptimizationBBO = "0.1"
+OptimizationOptimJL = "0.1"
 Plots = "1"
-
