+++ test/Project.toml
@@ -2,11 +2,11 @@
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
-GalacticBBO = "80c49c3a-6557-47d9-8f5b-13d0a2920315"
-GalacticOptim = "a75be94c-b780-496d-a8a9-0878b188d577"
-GalacticOptimJL = "9d3c5eb1-403b-401b-8c0f-c11105342e6b"
 MeasureBase = "fa1605e6-acd5-459c-a1e6-7e635759db14"
 Optim = "429524aa-4258-5aef-a3af-852621145aeb"
+Optimization = "7f7a1694-90dd-40f0-9382-eb1efda571ba"
+OptimizationBBO = "3e6eede4-6085-4f62-9a71-46d9bc1eb92b"
+OptimizationOptimJL = "36348300-93cb-4f02-beb5-3c3902f8871e"
 Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 PyCall = "438e738f-606a-5dbb-bf0a-cddfbfd45ab0"
