### 2022-07-14 bb4edf4
CompatHelper: add new compat entry for Comrade at version 0.4 for package docs, (keep existing compat)
