+++ docs/src/benchmarks.md
@@ -144,15 +144,93 @@ preh[2]["x0"] = Dict("prior_type"=>"flat", "min"=>-μas2rad(40.0), "max"=>μas2r
 preh[2]["y0"] = Dict("prior_type"=>"flat", "min"=>-μas2rad(40.0), "max"=>μas2rad(40.0))
 preh[2]["PA"] = Dict("prior_type"=>"flat", "min"=>-1π, "max"=>1π)
 
-# Now get the posterior function
-obj_func = 
+using PyCall
+py"""
+import ehtim
+import numpy as np
+transform_param = ehtim.modeling.modeling_utils.transform_param
+def make_paraminit(param_map, meh, trial_model, model_prior):
+    model_init = meh.copy()
+    param_init = []
+    for j in range(len(param_map)):
+        pm = param_map[j]
+        if param_map[j][1] in trial_model.params[param_map[j][0]].keys():
+            param_init.append(transform_param(model_init.params[pm[0]][pm[1]]/pm[2], model_prior[pm[0]][pm[1]],inverse=False))
+        else: # In this case, the parameter is a list of complex numbers, so the real/imaginary or abs/arg components need to be assigned
+            if param_map[j][1].find('cpol') != -1:
+                param_type = 'beta_list_cpol'
+                idx = int(param_map[j][1].split('_')[0][8:])
+            elif param_map[j][1].find('pol') != -1:
+                param_type = 'beta_list_pol'
+                idx = int(param_map[j][1].split('_')[0][7:]) + (len(trial_model.params[param_map[j][0]][param_type])-1)//2
+            elif param_map[j][1].find('beta') != -1:
+                param_type = 'beta_list'
+                idx = int(param_map[j][1].split('_')[0][4:]) - 1
+            else:
+                raise Exception('Unsure how to interpret ' + param_map[j][1])
 
+            curval = model_init.params[param_map[j][0]][param_type][idx]
+            if '_' not in param_map[j][1]:
+                param_init.append(transform_param(np.real( model_init.params[pm[0]][param_type][idx]/pm[2]), model_prior[pm[0]][pm[1]],inverse=False))
+            elif   param_map[j][1][-2:] == 're':
+                param_init.append(transform_param(np.real( model_init.params[pm[0]][param_type][idx]/pm[2]), model_prior[pm[0]][pm[1]],inverse=False))
+            elif param_map[j][1][-2:] == 'im':
+                param_init.append(transform_param(np.imag( model_init.params[pm[0]][param_type][idx]/pm[2]), model_prior[pm[0]][pm[1]],inverse=False))
+            elif param_map[j][1][-3:] == 'abs':
+                param_init.append(transform_param(np.abs(  model_init.params[pm[0]][param_type][idx]/pm[2]), model_prior[pm[0]][pm[1]],inverse=False))
+            elif param_map[j][1][-3:] == 'arg':
+                param_init.append(transform_param(np.angle(model_init.params[pm[0]][param_type][idx])/pm[2], model_prior[pm[0]][pm[1]],inverse=False))
+            else:
+                if not quiet: print('Parameter ' + param_map[j][1] + ' not understood!')
+    n_params = len(param_init)
+    return n_params, param_init
+"""
+
+# make the python param map and use optimize so we flatten the parameter space.
+pmap, pmask = ehtim.modeling.modeling_utils.make_param_map(meh, preh, "scipy.optimize.dual_annealing", fit_model=true)
+trial_model = meh.copy()
+
[…]
