+++ examples/imaging_closures.jl
@@ -1,7 +1,7 @@
 using Pkg; Pkg.activate(@__DIR__)
 using Comrade
 using Distributions
-using ComradeGalactic
+using ComradeOptimization
 using ComradeAHMC
 using GalacticBBO
 using Plots
@@ -41,15 +41,16 @@ function (model::ImModel)(θ)
     #Create the modelimage object that will use a cache to compute the DFT
     m = modelimage(img, cache)
 end
-const npix = 10
-const fovxy = μas2rad(65.0)
+
+npix = 32
+fovxy = μas2rad(65.0)
 # Now we can feed in the array information to form the cache. We will be using a DFT since
 # it is efficient for so few pixels
 cache = create_cache(Comrade.DFTAlg(dlcamp), IntensityMap(rand(npix,npix), fovxy, fovxy, BSplinePulse{3}()))
 mms = ImModel(cache, fovxy, npix)
 # We will use a Dirichlet prior to enforce that the flux sums to unity since closures are
 # degenerate to total flux.
-prior = (c = DistributionsAD.TuringDirichlet(npix*npix, 1.0),)
+prior = (c = MvNormal(fill(-5.0, npix^2), 1.0),)
 
 post = Posterior(lklhd, prior, mms)
 tpost = asflat(post)
@@ -58,7 +59,7 @@ tpost = asflat(post)
 # It turns out that gradients are really helpful here
 ndim = dimension(tpost)
 f = OptimizationFunction(tpost, GalacticOptim.AutoZygote())
-x0 = randn(ndim)
+x0 = rand(ndim)
 prob = OptimizationProblem(f, x0, nothing)
 sol = solve(prob, LBFGS(); maxiters=4_000, show_trace=true, show_every=50, g_tol=1e-2)
 
@@ -74,12 +75,210 @@ plot(mms(xopt), fovx=1.2*fovxy, fovy=1.2*fovxy)
 metric = DenseEuclideanMetric(ndim)
 hchain, stats = sample(post, AHMC(;metric, autodiff=AD.ZygoteBackend()), 3000; nadapts=2000, init_params=xopt)
 
+lca = logdensityof(lklhd.lklhds[1])
+lcp = logdensityof(lklhd.lklhds[2])
 
-foox = let dc=dc, dd=dd, pr=tpost.lpost.prior, tr=tpost.transform, plan = cache.plan, phases=cache.phases, dmat=dlcamp.config.designmat, dmatc=dcphase.config.designmat
+foox = let lca=lca, lcp=lcp, pr=tpost.lpost.prior, tr=tpost.transform, plan = cache.plan, phases=cache.phases, dmat=dlcamp.config.designmat, dmatc=dcphase.config.designmat
     x->begin
         y = transform(tr, x)
         lp = logdensityof(pr, y)
-        vis = conj.(plan*y.c).*phases
-        logdensityof(dd, dmat*log.(abs.(vis))) + logdensityof(dc, dmatc*angle.(vis)) + lp
+        vis = plan*exp.(y.c).*phases
+        return lca(vis) + lcp(vis) + lp
     end
 end
+
+
+
+
+function bench(npix, lklhd, dlcamp, )
+    fovxy = μas2rad(65.0)
+    # Now we can feed in the array information to form the cache. We will be using a DFT since
+    # it is efficient for so few pixels
[…]
