+++ examples/Project.toml
@@ -1,10 +1,11 @@
 [deps]
 AbstractDifferentiation = "c29ec348-61ec-40c8-8164-b8c60e9d9f3d"
+ChainRulesCore = "d360d2e6-b24c-11e9-a2a3-2a2ae2dbcce4"
 ComponentArrays = "b0b7db55-cfe3-40fc-9ded-d10e2dbeff66"
 Comrade = "99d987ce-9a1e-4df8-bc0b-1ea019aa547b"
 ComradeAHMC = "a4336a5c-78bc-4363-8a90-ce3fa9d3abe4"
 ComradeBase = "6d8c423b-a35f-4ef1-850c-862fe21f82c4"
-ComradeGalactic = "47701775-187b-49a4-b924-794639d71cdc"
+ComradeOptimization = "26988f03-70df-4135-959e-06268cc6415a"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 DistributionsAD = "ced4e74d-a319-5a8a-b0ac-84af2272839c"
@@ -13,13 +14,16 @@ ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
 GalacticBBO = "80c49c3a-6557-47d9-8f5b-13d0a2920315"
 GalacticOptimJL = "9d3c5eb1-403b-401b-8c0f-c11105342e6b"
 HypercubeTransform = "9ec9aee3-0fd3-44c2-8e61-a50acc66f3c8"
+Optimization = "7f7a1694-90dd-40f0-9382-eb1efda571ba"
 PairPlots = "43a3c2be-4208-490b-832a-a21dcd55d7da"
 Pathfinder = "b1d3bc72-d0e7-4279-b92f-7fa5d6d2d454"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 ReverseDiff = "37e2e3b7-166d-5795-8a7a-e32c996b4267"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
+StatsFuns = "4c63d2b9-4356-54db-8cca-17b64c39e42c"
 StatsPlots = "f3b207a7-027a-5e70-b257-86293d7955fd"
 TupleVectors = "615932cf-77b6-4358-adcd-5b7eba981d7e"
+Yota = "cd998857-8626-517d-b929-70ad188a48f0"
 ZigZagBoomerang = "36347407-b186-4a6a-8c98-4f4567861712"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
