+++ examples/paper_example.jl
@@ -1,74 +1,61 @@
-using Pkg; Pkg.activate(@__DIR__)
+using Pkg;Pkg.activate(@__DIR__)
 using Comrade
 using Distributions
-using ComradeOptimization
 using ComradeAHMC
-using GalacticBBO
-using GalacticOptimJL
+using ComradeOptimization
+using OptimizationBBO
 using Plots
-using StatsBase
-
 # load eht-imaging we use this to load eht data
 load_ehtim()
 # To download the data visit https://doi.org/10.25739/g85n-f134
-obs = ehtim.obsdata.load_uvfits(joinpath(@__DIR__, "SR1_M87_2017_096_lo_hops_netcal_StokesI.uvfits"))
+file = "SR1_M87_2017_096_lo_hops_netcal_StokesI.uvfits"
+obs = ehtim.obsdata.load_uvfits(joinpath(@__DIR__, file))
 obs.add_scans()
 # kill 0-baselines since we don't care about
-# large scale flux and make scan-average data
+# large scale structure and make scan-average data
 obs = obs.flag_uvdist(uv_min=0.1e9).avg_coherent(0.0, scan_avg=true)
 # extract log closure amplitudes and closure phases
 dlcamp = extract_lcamp(obs)
 dcphase = extract_cphase(obs)
 # form the likelihood
 lklhd = RadioLikelihood(dlcamp, dcphase)
-# build the model here we fit a ring with a azimuthal
-#brightness variation and a Gaussian
-function model(θ)
-    (;rad, wid, a, b, f, sig, asy, pa, x, y) = θ
-    ring = f*smoothed(stretched(MRing((a,), (b,)), μas2rad(rad), μas2rad(rad)), μas2rad(wid))
-    g = (1-f)*shifted(rotated(stretched(Gaussian(), μas2rad(sig)*asy, μas2rad(sig)), pa), μas2rad(x), μas2rad(y))
-    return ring + g
+# build the model: here we fit a ring with a azimuthal
+# brightness variation and a Gaussian
+function model(params)
+  (;rad, wid, a, b, f, sig, asy, pa, x, y) = params
+  ring = f*smoothed(stretched(MRing((a,), (b,)), rad, rad), wid)
+  g = (1-f)*shifted(rotated(stretched(Gaussian(), sig*asy, sig), pa), x, y)
+  return ring + g
 end
 # define the priors
+uas2rad = pi/180.0/3600/1e6
 prior = (
-          rad = Uniform(10.0, 30.0),
-          wid = Uniform(1.0, 10.0),
+          rad = Uniform(uas2rad*(10.0), uas2rad*(30.0)),
+          wid = Uniform(uas2rad*(1.0), uas2rad*(10.0)),
           a = Uniform(-0.5, 0.5), b = Uniform(-0.5, 0.5),
           f = Uniform(0.0, 1.0),
-          sig = Uniform((1.0), (60.0)),
-          asy = Uniform(0.0, 0.9),
-          pa = Uniform(0.0, 1π),
-          x = Uniform(-(80.0), (80.0)),
-          y = Uniform(-(80.0), (80.0))
+          sig = Uniform(uas2rad*(1.0), uas2rad*(40.0)),
+          asy = Uniform(0.0, 0.75),
+          pa = Uniform(0.0, 1pi),
+          x = Uniform(-uas2rad*(80.0), uas2rad*(80.0)),
+          y = Uniform(-uas2rad*(80.0), uas2rad*(80.0))
         )
 # Now form the posterior
 post = Posterior(lklhd, prior, model)
 # We will use HMC to sample the posterior.
-# First we will find a reasonable starting location using GalacticOptim
-# For optimization we need to specify what transform to use. Here we will transform to
-# the unit hypercube
-tpost = asflat(post)
[…]
