+++ lib/ComradeNested/Project.toml
@@ -1,7 +1,7 @@
 name = "ComradeNested"
 uuid = "233675ba-9426-43ce-b06d-28ea4529d7f1"
 authors = ["Paul Tiede <ptiede91@gmail.com>"]
-version = "0.1.0"
+version = "0.2.0"
 
 [deps]
 AbstractMCMC = "80f14c24-f653-4e6a-9b94-39d6b0f70001"
@@ -12,11 +12,11 @@ TupleVectors = "615932cf-77b6-4358-adcd-5b7eba981d7e"
 
 [compat]
 AbstractMCMC = "3, 4"
-Comrade = "0.3,0.4"
+Comrade = "0.5"
 NestedSamplers = "0.8"
 Reexport = "1"
 TupleVectors = "0.1, 0.2"
-julia = "1.6"
+julia = "1.7"
 
 [extras]
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
