### 2022-07-15 510fde0
fix uuid in ComradeOptimization test
