+++ lib/ComradeOptimization/Project.toml
@@ -24,7 +24,7 @@ julia = "1.7"
 [extras]
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 OptimizationBBO = "3e6eede4-6085-4f62-9a71-46d9bc1eb92b"
-OptimizationOptimJL = "9d3c5eb1-403b-401b-8c0f-c11105342e6b"
+OptimizationOptimJL = "36348300-93cb-4f02-beb5-3c3902f8871e"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
