### 2022-07-16 77786dd
CompatHelper: add new compat entry for Comrade at version 0.5 for package docs, (keep existing compat)
