+++ lib/ComradeOptimization/Project.toml
@@ -13,7 +13,7 @@ Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 SciMLBase = "0bca4576-84f4-4d90-8ffe-ffa030f20462"
 
 [compat]
-Comrade = "0.5"
+Comrade = "0.5, 0.6"
 Distributions = "0.24,0.25"
 ForwardDiff = "0.9, 0.10"
 Optimization = "3"
