+++ lib/ComradeAHMC/Project.toml
@@ -1,7 +1,7 @@
 name = "ComradeAHMC"
 uuid = "a4336a5c-78bc-4363-8a90-ce3fa9d3abe4"
 authors = ["Paul Tiede <ptiede91@gmail.com> and contributors"]
-version = "0.2.0"
+version = "0.2.1"
 
 [deps]
 AbstractDifferentiation = "c29ec348-61ec-40c8-8164-b8c60e9d9f3d"
@@ -18,7 +18,7 @@ AbstractDifferentiation = "0.4"
 AbstractMCMC = "3,4"
 AdvancedHMC = "0.3"
 ArgCheck = "2"
-Comrade = "0.5"
+Comrade = "0.5, 0.6"
 Reexport = "1"
 TupleVectors = "0.1, 0.2"
 julia = "1.7"
