+++ lib/ComradeAHMC/Project.toml
@@ -26,6 +26,7 @@ julia = "1.7"
 [extras]
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [targets]
-test = ["Test", "Distributions"]
+test = ["Test", "Distributions", "Zygote"]
