+++ docs/Project.toml
@@ -22,7 +22,7 @@ Distributions = "0.25"
 Documenter = "0.27"
 ForwardDiff = "0.10"
 Glob = "1"
-HypercubeTransform = "^0.2.7"
+HypercubeTransform = "0.3"
 Literate = "2"
 Optimization = "3"
 OptimizationBBO = "0.1"
