### 2022-07-20 69973fb
include dynesty in docs ci
