### 2022-07-20 001fedf
CompatHelper: add new compat entry for ComradeDynesty at version 0.2 for package docs, (keep existing compat)
