### 2022-07-20 705c275
CompatHelper: add new compat entry for ComradeNested at version 0.2 for package docs, (keep existing compat)
