+++ docs/Project.toml
@@ -20,6 +20,7 @@ Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 [compat]
 ComradeAHMC = "0.1,0.2"
 ComradeBase = "0.3"
+ComradeNested = "0.2"
 ComradeOptimization = "0.1"
 Distributions = "0.25"
 Documenter = "0.27"
