### 2022-07-20 cc4fa93
CompatHelper: add new compat entry for ComradeAdaptMCMC at version 0.2 for package docs, (keep existing compat)
