### 2022-07-20 4b1d12b
CompatHelper: add new compat entry for Comrade at version 0.6 for package docs, (keep existing compat)
