+++ Project.toml
@@ -73,7 +73,7 @@ HypercubeTransform = "0.3"
 KeywordCalls = "0.2"
 MeasureBase = "0.7, 0.8, 0.9"
 MeasureTheory = "0.15, 0.16"
-NFFT = "0.10, 0.11, 0.12"
+NFFT = "0.10, 0.11, 0.12, 0.13"
 NamedTupleTools = "0.13, 0.14"
 PaddedViews = "0.5"
 ParameterHandling = "0.4"
