### 2022-07-21 e66abf6
update the examples to work with the new version
