+++ examples/imaging_closures.jl
@@ -1,13 +1,14 @@
 using Pkg; Pkg.activate(@__DIR__)
+Pkg.add("https://github.com/ptiede/RadioImagePriors.jl")
 using Comrade
 using Distributions
 using ComradeOptimization
 using ComradeAHMC
-using GalacticBBO
+using OptimizationBBO
 using Plots
 using StatsBase
-using GalacticOptimJL
-using DistributionsAD
+using OptimizationOptimJL
+using RadioImagePriors
 
 # load eht-imaging we use this to load eht data
 load_ehtim()
@@ -28,6 +29,12 @@ struct ImModel{C}
     cache::C
     fov::Float64
     npix::Int
+    function ImModel(obs::Comrade.EHTObservation, fov::Real, npix::Int)
+        buffer = IntensityMap(zeros(npix, npix), fov, fov, BSplinePulse{3}())
+        cache = create_cache(DFTAlg(obs), buffer)
+        return new{typeof(cache)}(cache, fov, npix)
+    end
+
 end
 
 # For our model we will be using a rasterized image. This can be viewed as something like a
@@ -36,21 +43,21 @@ end
 function (model::ImModel)(θ)
     (;c) = θ
     #Construct the image model
-    cmat = reshape(c, model.npix, model.npix)
-    img = IntensityMap(cmat, model.fov, model.fov, BSplinePulse{3}())
+    img = IntensityMap(c, model.fov, model.fov, BSplinePulse{3}())
     #Create the modelimage object that will use a cache to compute the DFT
     m = modelimage(img, cache)
 end
 
-npix = 32
-fovxy = μas2rad(65.0)
+
+npix = 24
+fovxy = μas2rad(62.5)
 # Now we can feed in the array information to form the cache. We will be using a DFT since
 # it is efficient for so few pixels
 cache = create_cache(Comrade.DFTAlg(dlcamp), IntensityMap(rand(npix,npix), fovxy, fovxy, BSplinePulse{3}()))
 mms = ImModel(cache, fovxy, npix)
 # We will use a Dirichlet prior to enforce that the flux sums to unity since closures are
 # degenerate to total flux.
-prior = (c = MvNormal(fill(-5.0, npix^2), 1.0),)
+prior = (c = ImageDirichlet(0.5, npix, npix),)
 
 post = Posterior(lklhd, prior, mms)
 tpost = asflat(post)
@@ -58,227 +65,44 @@ tpost = asflat(post)
 # Let's run an optimizer to get a nice starting location
 # It turns out that gradients are really helpful here
 ndim = dimension(tpost)
-f = OptimizationFunction(tpost, GalacticOptim.AutoZygote())
-x0 = rand(ndim)
-prob = OptimizationProblem(f, x0, nothing)
-sol = solve(prob, LBFGS(); maxiters=4_000, show_trace=true, show_every=50, g_tol=1e-2)
+using Zygote
+f = OptimizationFunction(tpost, Optimization.AutoZygote())
+prob = OptimizationProblem(f, randn(ndim), nothing)
+ℓ = logdensityof(tpost)
+sol = solve(prob, LBFGS(); maxiters=2_000, callback=(x,p)->(@info ℓ(x); false), g_tol=1e-1)
 
 xopt = transform(tpost, sol)
 
 # Let's see how the fit looks
[…]
