+++ examples/benchmarks.jl
@@ -1,0 +1,169 @@
+using Comrade
+using Distributions
+
+load_ehtim()
+# To download the data visit https://doi.org/10.25739/g85n-f134
+obs = ehtim.obsdata.load_uvfits(joinpath(@__DIR__, "SR1_M87_2017_096_lo_hops_netcal_StokesI.uvfits"))
+obs.add_scans()
+obsavg = obs.avg_coherent(0.0, scan_avg=true)
+amp = extract_amp(obsavg)
+lklhd = RadioLikelihood(amp)
+
+function model(θ)
+    (;rad, wid, a, b, f, sig, asy, pa, x, y) = θ
+    ring = f*smoothed(stretched(MRing((a,), (b,)), μas2rad(rad), μas2rad(rad)), μas2rad(wid))
+    g = (1-f)*shifted(rotated(stretched(Gaussian(), μas2rad(sig)*asy, μas2rad(sig)), pa), μas2rad(x), μas2rad(y))
+    return ring + g
+end
+prior = (
+          rad = Uniform(10.0, 30.0),
+          wid = Uniform(1.0, 10.0),
+          a = Uniform(-0.5, 0.5), b = Uniform(-0.5, 0.5),
+          f = Uniform(0.0, 1.0),
+          sig = Uniform((1.0), (60.0)),
+          asy = Uniform(0.0, 0.9),
+          pa = Uniform(0.0, 1π),
+          x = Uniform(-(80.0), (80.0)),
+          y = Uniform(-(80.0), (80.0))
+        )
+# Now form the posterior
+post = Posterior(lklhd, prior, model)
+
+θ = (f=0.8, rad= 22.0, wid= 3.0, a = 0.0, b = 0.15, sig = 20.0, asy=0.2, pa=π/2, x=20.0, y=20.0)
+m = model(θ)
+plot(m)
+
+post = Posterior(lklhd, prior, model)
+tpost = asflat(post)
+
+x0 = inverse(tpost, θ)
+
+ℓ = logdensityof(tpost)
+@benchmark ℓ($x0)
+
+using ForwardDiff
+gℓ = Base.Fix1(ForwardDiff.gradient, ℓ)
+@benchmark gℓ($x0)
+
+meh = ehtim.model.Model()
+meh = meh.add_thick_mring(F0=θ.f,
+                    d=2*μas2rad(θ.rad),
+                    alpha=2*sqrt(2*log(2))*μas2rad(θ.wid),
+                    x0 = 0.0,
+                    y0 = 0.0,
+                    beta_list=[0.0+θ.b]
+                    )
+meh = meh.add_gauss(F0=1-θ.f,
+                    FWHM_maj=2*sqrt(2*log(2))*μas2rad(θ.sig),
+                    FWHM_min=2*sqrt(2*log(2))*μas2rad(θ.sig)*θ.asy,
+                    PA = θ.pa,
+                    x0 = μas2rad(20.0),
+                    y0 = μas2rad(20.0)
+                    )
+
+preh = meh.default_prior()
+preh[1]["F0"] = Dict("prior_type"=>"flat", "min"=>0.0, "max"=>1.0)
+preh[1]["d"] = Dict("prior_type"=>"flat", "min"=>μas2rad(20.0), "max"=>μas2rad(60.0))
+preh[1]["alpha"] = Dict("prior_type"=>"flat", "min"=>μas2rad(2.0), "max"=>μas2rad(25.0))
+preh[1]["x0"] = Dict("prior_type"=>"fixed")
+preh[1]["y0"] = Dict("prior_type"=>"fixed")
+
[…]
