### 2022-07-21 3bc2178
clean up _visibilities
