### 2022-07-21 bb053a2
add some inlines
