### 2022-07-21 3c2e6c8
Fix up imaging closures
