### 2022-07-21 dc15761
fix global cache bug
