+++ examples/imaging_closures.jl
@@ -1,5 +1,5 @@
 using Pkg; Pkg.activate(@__DIR__)
-Pkg.add("https://github.com/ptiede/RadioImagePriors.jl")
+Pkg.add(url="https://github.com/ptiede/RadioImagePriors.jl")
 using Comrade
 using Distributions
 using ComradeOptimization
@@ -45,7 +45,7 @@ function (model::ImModel)(θ)
     #Construct the image model
     img = IntensityMap(c, model.fov, model.fov, BSplinePulse{3}())
     #Create the modelimage object that will use a cache to compute the DFT
-    m = modelimage(img, cache)
+    m = modelimage(img, model.cache)
 end
 
 
