### 2022-07-22 b0ff2a9
Update imaging_closures.jl
