+++ examples/imaging.jl
@@ -1,15 +1,14 @@
-
-
 using Pkg; Pkg.activate(@__DIR__)
+Pkg.add(url="https://github.com/ptiede/RadioImagePriors.jl")
 using Comrade
 using Distributions
-using ComradeAHMC
-using Plots
-using Zygote
 using ComradeOptimization
+using ComradeAHMC
 using OptimizationBBO
-using Optim
-using NLopt
+using Plots
+using StatsBase
+using OptimizationOptimJL
+using RadioImagePriors
 
 # load eht-imaging we use this to load eht data
 load_ehtim()
@@ -20,79 +19,124 @@ obs.add_scans()
 # large scale flux and make scan-average data
 obs = obs.flag_uvdist(uv_min=0.1e9).avg_coherent(0.0, scan_avg=true).add_fractional_noise(0.02)
 # extract log closure amplitudes and closure phases
-vis = extract_vis(obs)
-lklhd = RadioLikelihood(vis)
+damp = extract_amp(obs)
+dcphase = extract_cphase(obs)
+lklhd = RadioLikelihood(damp, dcphase)
 
 # Build the Model. Here we we a struct to hold some caches
 # This will be useful to hold precomputed caches
 
-struct Model{C,G}
+struct GModel{C,G}
     cache::C
     gcache::G
     fov::Float64
     npix::Int
+    function GModel(obs::Comrade.EHTObservation, fov, npix)
+        buffer = IntensityMap(zeros(npix, npix), fov, fov, BSplinePulse{3}())
+        cache = create_cache(DFTAlg(obs), buffer)
+        gcache = GainCache(scantable(obs))
+        return new{typeof(cache), typeof(gcache)}(cache, gcache, fov, npix)
+    end
 end
 
-function (model::Model)(θ)
-    ;c, f, gamp, gphase = θ
+function (model::GModel)(θ)
+    (;c, f, lgamp) = θ
     # Construct the image model
-    cmat = f.*reshape(c, model.npix, model.npix)
-    img = IntensityMap(cmat, model.fov, model.fov, BSplinePulse{3}())
-    m = modelimage(img, cache)
+    img = IntensityMap(f*c, model.fov, model.fov, BSplinePulse{3}())
+    m = modelimage(img, model.cache)
     # Now corrupt the model with Gains
-    g = gamp.*cis.(gphase)
-    Comrade.GainModel(gcache, g, m)
+    g = exp.(lgamp)
+    Comrade.GainModel(model.gcache, g, m)
 end
 
 # First we define the station gain priors
-distamp = (AA = LogNormal(0.0, 0.1),
-         AP = LogNormal(0.0, 0.1),
-         LM = LogNormal(0.0, 0.5),
-         AZ = LogNormal(0.0, 0.1),
-         JC = LogNormal(0.0, 0.1),
-         PV = LogNormal(0.0, 0.1),
-         SM = LogNormal(0.0, 0.1))
-# We will use AA as a reference array
-distphase = (AA = VonMises(0.0, 10000.0),
-             AP = VonMises(0.0, 1/π^2),
-             LM = VonMises(0.0, 1/π^2),
-             AZ = VonMises(0.0, 1/π^2),
-             JC = VonMises(0.0, 1/π^2),
-             PV = VonMises(0.0, 1/π^2),
-             SM = VonMises(0.0, 1/π^2))
+distamp = (AA = Normal(0.0, 0.1),
+           AP = Normal(0.0, 0.1),
+           LM = Normal(0.0, 0.9),
[…]
