+++ src/models/geometric_models.jl
@@ -165,6 +165,13 @@ function MRing(a::Real, b::Real)
     return MRing([aT], [bT])
 end
 
+function MRing(a::Tuple, b::Tuple)
+    return MRing(collect(a), collect(b))
+end
+
+# Depreciate this method since we are moving to vectors for simplicity.
+@deprecate MRing(a::Tuple, b::Tuple) MRing(a::AbstractVector, b::AbstractVector)
+
 radialextent(::MRing) = 1.5
 
 
