### 2022-07-31 e22b536
fix benchmark example script
