+++ test/Core/models.jl
@@ -99,8 +99,8 @@ end
         test_rrule(Comrade.visibility_point, MRing(α, β), 0.5, 0.25)
         # We convolve it to remove some pixel effects
         m = convolved(MRing(α, β), stretched(Gaussian(), 0.1, 0.1))
-        m2 = convolved(MRing{1}(collect(α), collect(β)), stretched(Gaussian(), 0.1, 0.1))
-        @test m == m2
+        m2 = convolved(MRing(α[1], β[1]), stretched(Gaussian(), 0.1, 0.1))
+        @test visibility(m, 0.1, 0.1) == visibility(m2, 0.1, 0.1)
         testmodel(m, 2048, 1e-3)
     end
 
