### 2022-07-31 e05cec9
update black hole image examples
