+++ docs/src/vlbi_imaging_problem.md
@@ -17,13 +17,14 @@ The goal of `Comrade` is to calculate (or approximate) the posterior of a image 
 
 ## The Likelihood
 
-`Comrade` uses the traditional VLBI expression for the data likelihood. Following [^TMS], each complex visibility ``V_{ij}`` follows a complex normal distribution
+`Comrade` uses the traditional VLBI expression for the data likelihood. Following standard VLBI data analysis[^TMS] each complex visibility ``V_{ij}`` follows a complex normal distribution
 
 ```math 
 p(V_{ij} | I) = (2\pi \sigma^2_{ij})^{-1/2}\exp\left(-\frac{| V_{ij} - \hat{V}_{ij}|^2}{2\sigma^2_{ij}}\right),
 ```
+where ``\hat{V}`` denotes the model visibility that we are comparing to the data. This likelihood forms the basis for all data analysis. The modeled $\hat{V}$ depend specifically on the model chosen. As a result `Comrade` has implemented a large number of models to describe observations. For a list of potential models please see [Comrade API](@ref).
 
-For other related data products, e.g. visibility amplitudes, closure products, we use the high-signal-to-noise Gaussian approximations of the distributions. For the specific expressions please see [^SgrP4].
+For other related data products, e.g. visibility amplitudes, closure products, we use the high-signal-to-noise Gaussian approximations of the distributions. For the specific expressions please see EHTC 2022 Paper IV[^SgrP4].
 
 [^TMS]: Thompson, A., Moran, J., Swenson, G. (2017). Interferometry and Synthesis in Radio Astronomy (Third). Springer Cham
 [^SgrP4]: Event Horizon Telescope Collaboration, (2022). First Sagittarius A* Event Horizon Telscope Results. IV. Variability, Morphology, and Black Hole Mass. ApJL 930 L15 [doi](10.3847/2041-8213/ac667410.3847/2041-8213/ac667210.3847/2041-8213/ac6736)
