### 2022-08-01 be54a49
add MCMC diagnostics
