### 2022-08-01 b182552
CompatHelper: add new compat entry for MCMCDiagnostics at version 0.3 for package docs, (keep existing compat)
