+++ docs/Project.toml
@@ -31,6 +31,7 @@ ForwardDiff = "0.10"
 Glob = "1"
 HypercubeTransform = "0.3"
 Literate = "2"
+MCMCDiagnostics = "0.3"
 Optimization = "3"
 OptimizationBBO = "0.1"
 OptimizationOptimJL = "0.1"
