### 2022-08-01 408fe74
CompatHelper: add new compat entry for Comrade at version 0.6 for package docs, (keep existing compat)
