### 2022-08-02 1fb9a65
update the docs to include a description of the imaging problem.
