+++ docs/src/vlbi_imaging_problem.md
@@ -12,20 +12,54 @@ p(I|V) = \frac{p(V|I)p(I)}{p(V)}.
 
 Here ``p(V|I)`` is known as the likelihood and describes the probability distribution of the data given some image `I`. The prior ``p(I)`` encodes prior knowledge of the image structure. This prior includes distributions of model parameters and even the model itself. Finally, the denominator ``p(V)`` is a normalization term and is known as the marginal likelihood or evidence and can be used to assess how well particular models fit the data.
 
-The goal of `Comrade` is to calculate (or approximate) the posterior of a image model. To see how to do that please see the [Making an Image of a Black Hole](@ref) tutorial.
+Therefore, to calculate our posterior we need to first specify the likelihood and prior. Below we provide a brief description of the likelihoods and models/priors that `Comrade` uses. However, if the user just wants to see how everything works first, they should check out the [Making an Image of a Black Hole](@ref) tutorial.
 
+## Likelihood
 
-## The Likelihood
-
-`Comrade` uses the traditional VLBI expression for the data likelihood. Following standard VLBI data analysis[^TMS] each complex visibility ``V_{ij}`` follows a complex normal distribution
+Following TMS[^TMS], we note that the likelihood for a single complex visibility at baseline ``u_{ij}, v_{ij}`` is
 
 ```math 
-p(V_{ij} | I) = (2\pi \sigma^2_{ij})^{-1/2}\exp\left(-\frac{| V_{ij} - \hat{V}_{ij}|^2}{2\sigma^2_{ij}}\right),
+p(V_{ij} | I) = (2\pi \sigma^2_{ij})^{-1/2}\exp\left(-\frac{| V_{ij} - g_ig_j^*\tilde{I}_{ij}(I)|^2}{2\sigma^2_{ij}}\right).
 ```
-where ``\hat{V}`` denotes the model visibility that we are comparing to the data. This likelihood forms the basis for all data analysis. The modeled $\hat{V}$ depend specifically on the model chosen. As a result `Comrade` has implemented a large number of models to describe observations. For a list of potential models please see [Comrade API](@ref).
 
-For other related data products, e.g. visibility amplitudes, closure products, we use the high-signal-to-noise Gaussian approximations of the distributions. For the specific expressions please see EHTC 2022 Paper IV[^SgrP4].
+In this equation ``\tilde{I}`` is the Fourier transform of the image ``I``, and ``g_{i,j}`` are complex numbers known as gains. The gains arise due to atmospheric and telescope effects and corrupt the incoming signal. Therefore, if a user attempts to model the complex visibilities they must also model the complex gains. An example showing how to model gains in `Comrade` can be found in [`examples/imaging.jl`](https://github.com/ptiede/Comrade.jl/tree/main/examples) on GitHub.
+
+Modeling the gains however can be computationally expensive, especially if our image model is simple. For instance in `Comrade` we have a wide variety of geometric models. These models tend to have a small number of parameters, and are simple to evaluate. In this instance simulataneously solving for gains is extremely expensive. As a result part of the EHT analysis[^M87P6][^SgrAP4] instead uses closure products as the data to fit. The two forms of closure products are:
+
+  - Closure Phases,
+  - Log-Closure Amplitudes.
+
+Closure Phases ``\psi`` are constructed by selecting three baselines ``(i,j,k)`` and finding the argument of the bispectrum
+
+```math
+    \psi_{ijk} = \arg V_{ij}V_{jk}V_{ki}.
+```
+
+Similar log-closure amplitudes are found by selecting four baselines ``(i,j,k,l)`` and forming the closure amplitudes
+
+```math
+    A_{ijkl} = \frac{ |V_{ij}||V_{kl}|}{|V_{jk}||V_{li}|}.
+```
+
[…]
