+++ docs/src/vlbi_imaging_problem.md
@@ -24,26 +24,28 @@ p(V_{ij} | I) = (2\pi \sigma^2_{ij})^{-1/2}\exp\left(-\frac{| V_{ij} - g_ig_j^*\
 
 In this equation ``\tilde{I}`` is the Fourier transform of the image ``I``, and ``g_{i,j}`` are complex numbers known as gains. The gains arise due to atmospheric and telescope effects and corrupt the incoming signal. Therefore, if a user attempts to model the complex visibilities they must also model the complex gains. An example showing how to model gains in `Comrade` can be found in [`examples/imaging.jl`](https://github.com/ptiede/Comrade.jl/tree/main/examples) on GitHub.
 
-Modeling the gains however can be computationally expensive, especially if our image model is simple. For instance in `Comrade` we have a wide variety of geometric models. These models tend to have a small number of parameters, and are simple to evaluate. In this instance simulataneously solving for gains is extremely expensive. As a result part of the EHT analysis[^M87P6][^SgrAP4] instead uses closure products as the data to fit. The two forms of closure products are:
+Modeling the gains can be computationally expensive, especially if our image model is simple. For instance in `Comrade` we have a wide variety of geometric models. These models tend to have a small number of parameters, and are simple to evaluate. Solving for gains then drastically increases the amount of time it takes to sample the posterior. As a result, part of the typical EHT analysis[^M87P6][^SgrAP4] instead uses closure products as its data. The two forms of closure products are:
 
   - Closure Phases,
   - Log-Closure Amplitudes.
 
-Closure Phases ``\psi`` are constructed by selecting three baselines ``(i,j,k)`` and finding the argument of the bispectrum
+Closure Phases ``\psi`` are constructed by selecting three baselines ``(i,j,k)`` and finding the argument of the bispectrum:
 
 ```math
     \psi_{ijk} = \arg V_{ij}V_{jk}V_{ki}.
 ```
 
-Similar log-closure amplitudes are found by selecting four baselines ``(i,j,k,l)`` and forming the closure amplitudes
+Similar log-closure amplitudes are found by selecting four baselines ``(i,j,k,l)`` and forming the closure amplitudes:
 
 ```math
     A_{ijkl} = \frac{ |V_{ij}||V_{kl}|}{|V_{jk}||V_{li}|}.
 ```
 
-Now instead of fitting closure amplitudes directly, it turns out that the statistically better behaved data quantities are the log-closure amplitudes. The benefit of fitting closure amplitudes is that they are independent of complex gains so we do not 
-have include them when modeling the data. However, the downside is that they effectively put uniform improper priors on the gains[^Blackburn] meaning that we are often throwing away information about the telescope performance. The other downside is that the 
-likelihoods for the closure products are not as simple. In the high-signal-to-noise-limit however they do reduce to Gaussian likelihoods. For the explicit likelihoods `Comrade` uses, we refer the reader to paper IV in the Sgr A* EHT publications[^SgrAP4].
+Now instead of fitting closure amplitudes directly, it turns out that the statistically better behaved data product is the log-closure amplitude. 
+
+The benefit of fitting closure products are that they are independent of complex gains so we do not 
+have include them when modeling the data. However, the downside is that they effectively put uniform improper priors on the gains[^Blackburn] meaning that we are often throwing away information about the telescope performance. On the other hand we can then view closure fitting as a very conservative estimate
[…]
