### 2022-08-02 bb9a52b
improve caltable docs
