### 2022-08-02 e8b6af2
update gain tests
