### 2022-08-02 6b55188
add nifty to paper
