+++ paper/paper.bib
@@ -8,6 +8,26 @@
       doi = {10.1137/141000671}
 }
 
+@ARTICLE{nifty,
+       author = {{Selig}, M. and {Bell}, M.~R. and {Junklewitz}, H. and {Oppermann}, N. and {Reinecke}, M. and {Greiner}, M. and {Pachajoa}, C. and {En{\ss}lin}, T.~A.},
+        title = "{NIFTY - Numerical Information Field Theory. A versatile PYTHON library for signal inference}",
+      journal = {\aap},
+     keywords = {methods: data analysis, methods: numerical, methods: statistical, techniques: image processing, Astrophysics - Instrumentation and Methods for Astrophysics, Computer Science - Information Theory, Computer Science - Mathematical Software, Mathematical Physics, Physics - Data Analysis, Statistics and Probability, Statistics - Computation},
+         year = 2013,
+        month = jun,
+       volume = {554},
+          eid = {A26},
+        pages = {A26},
+          doi = {10.1051/0004-6361/201321236},
+archivePrefix = {arXiv},
+       eprint = {1301.4499},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2013A&A...554A..26S},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
+
 @article{juliafast,
       author = {Bezanson, Jeff and Chen, Jiahao and Chung, Benjamin and Karpinski, Stefan and Shah, Viral B. and Vitek, Jan and Zoubritzky, Lionel},
       title = {Julia: Dynamism and Performance Reconciled by Design},
@@ -174,6 +194,25 @@ archivePrefix = {arXiv},
   doi          = {10.5281/zenodo.5808196},
   url          = {https://doi.org/10.5281/zenodo.5808196}
 }
+
+@ARTICLE{Arras,
+       author = {{Arras}, Philipp and {Frank}, Philipp and {Haim}, Philipp and {Knollm{\"u}ller}, Jakob and {Leike}, Reimar and {Reinecke}, Martin and {En{\ss}lin}, Torsten},
+        title = "{Variable structures in M87* from space, time and frequency resolved interferometry}",
+      journal = {Nature Astronomy},
+     keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Astrophysics of Galaxies},
+         year = 2022,
+        month = jan,
+       volume = {6},
+        pages = {259-269},
+          doi = {10.1038/s41550-021-01548-0},
+archivePrefix = {arXiv},
+       eprint = {2002.05218},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2022NatAs...6..259A},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
 
 @ARTICLE{EHTCIV,
        author = {{Event Horizon Telescope Collaboration}},
