### 2022-08-02 310d3fc
move abstractmccm to typedtables
