+++ lib/ComradeAdaptMCMC/Project.toml
@@ -1,19 +1,19 @@
 name = "ComradeAdaptMCMC"
 uuid = "5a6518a2-e6ab-4020-97b8-110ebf16b849"
 authors = ["Paul Tiede <ptiede91@gmail.com>"]
-version = "0.2.1"
+version = "0.2.2"
 
 [deps]
 AbstractMCMC = "80f14c24-f653-4e6a-9b94-39d6b0f70001"
 AdaptiveMCMC = "717c3277-546c-407d-8270-09a39a0919a0"
 Comrade = "99d987ce-9a1e-4df8-bc0b-1ea019aa547b"
-TupleVectors = "615932cf-77b6-4358-adcd-5b7eba981d7e"
+TypedTables = "9d95f2ec-7b3d-5a63-8d20-e2491e220bb9"
 
 [compat]
 AbstractMCMC = "3, 4"
 AdaptiveMCMC = "0.1"
 Comrade = "0.5, 0.6"
-TupleVectors = "0.1, 0.2"
+TypedTables = "1"
 julia = "1.7"
 
 [extras]
