+++ Project.toml
@@ -47,7 +47,6 @@ StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 StructArrays = "09ab397b-f2b6-538f-b94a-2f83cf4a842a"
 Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 TransformVariables = "84d833dd-6860-57f9-a1a7-6da5db126cff"
-TupleVectors = "615932cf-77b6-4358-adcd-5b7eba981d7e"
 UUIDs = "cf7118a7-6976-5b1a-9a39-7adc72f591a4"
 ValueShapes = "136a8f8c-c49b-4edb-8b98-f3d64d48be8f"
 
@@ -89,7 +88,6 @@ StatsBase = "0.31,0.32,0.33"
 StructArrays = "0.3,0.4,0.5,0.6"
 Tables = "1"
 TransformVariables = "0.5,0.6"
-TupleVectors = "0.1, 0.2"
 ValueShapes = "0.10"
 julia = "^1.6"
 
