+++ lib/ComradeDynesty/Project.toml
@@ -1,21 +1,21 @@
 name = "ComradeDynesty"
 uuid = "0e22d245-d796-46d0-bd67-132aa2ceb697"
 authors = ["Paul Tiede <ptiede91@gmail.com>"]
-version = "0.2.1"
+version = "0.2.2"
 
 [deps]
 AbstractMCMC = "80f14c24-f653-4e6a-9b94-39d6b0f70001"
 Comrade = "99d987ce-9a1e-4df8-bc0b-1ea019aa547b"
 Dynesty = "eb527566-0f3e-4aab-bb5f-9d2e403dba70"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
-TupleVectors = "615932cf-77b6-4358-adcd-5b7eba981d7e"
+TypedTables = "9d95f2ec-7b3d-5a63-8d20-e2491e220bb9"
 
 [compat]
 AbstractMCMC = "3, 4"
 Comrade = "0.5, 0.6"
 Dynesty = "0.1"
 Reexport = "1"
-TupleVectors = "0.1, 0.2"
+TypedTables = "1"
 julia = "1.7"
 
 [extras]
