### 2022-08-03 a87ecf3
update paper examples
