+++ examples/Project.toml
@@ -14,6 +14,7 @@ ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
 HypercubeTransform = "9ec9aee3-0fd3-44c2-8e61-a50acc66f3c8"
 MCMCDiagnosticTools = "be115224-59cd-429b-ad48-344e309966f0"
 MCMCDiagnostics = "6e857e4b-079a-58c4-aeab-bc2670384359"
+Measurements = "eff96d63-e80a-5855-80a2-b1b0885c5ab7"
 OptimizationBBO = "3e6eede4-6085-4f62-9a71-46d9bc1eb92b"
 OptimizationOptimJL = "36348300-93cb-4f02-beb5-3c3902f8871e"
 PairPlots = "43a3c2be-4208-490b-832a-a21dcd55d7da"
@@ -21,7 +22,8 @@ Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 RadioImagePriors = "b1ba175b-8447-452c-b961-7db2d6f7a029"
 ReverseDiff = "37e2e3b7-166d-5795-8a7a-e32c996b4267"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
-TupleVectors = "615932cf-77b6-4358-adcd-5b7eba981d7e"
+Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
+TypedTables = "9d95f2ec-7b3d-5a63-8d20-e2491e220bb9"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [extras]
