### 2022-08-03 09183db
update docs to use Tables Interface
