+++ docs/src/examples/black_hole_image.jl
@@ -163,18 +163,17 @@ residual(model(xopt), dlcamp)
 
 # For a real run we should also check that the MCMC chain has converged. For
 # this we can use MCMCDiagnostics
-using MCMCDiagnostics
+using MCMCDiagnostics, Tables
 # First lets look at the effective sample size or ESS. This is important since
 # MCMC estimates converges as √ESS (for most problems).
-ess = [r=effective_sample_size(getproperty(chain, r)) for r in propertynames(chain)]
+ess = map(effective_sample_size, Tables.columns(chain))
 # We can also calculate the split-rhat or potential scale reduction. For this we should actually
 # use at least 4 chains. However for demonstation purposes we will use one chain that we split in two
-rhats = NamedTuple{tuple(propertynames(chain))[1]}(map(propertynames(chain)) do n
-    c = getproperty(chain, n)
+rhats = map(Tables.columns(chain)) do c
     c1 = @view c[1001:1500]
-    c2 = @view c[1500:2000]
+    c2 = @view c[1501:2000]
     return potential_scale_reduction(c1, c2)
-end)
+end
 # Ok we have a split-rhat < 1.01 on all parameters so we have success (in reality run more chains!).
 
 
