+++ examples/Project.toml
@@ -8,6 +8,7 @@ ComradeBase = "6d8c423b-a35f-4ef1-850c-862fe21f82c4"
 ComradeOptimization = "26988f03-70df-4135-959e-06268cc6415a"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
+DistributionsAD = "ced4e74d-a319-5a8a-b0ac-84af2272839c"
 Enzyme = "7da242da-08ed-463a-9acd-ee780be4f1d9"
 FiniteDifferences = "26cc04aa-876d-5657-8c51-4c34ba976000"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
