### 2022-08-04 12ade6f
CompatHelper: add new compat entry for Comrade at version 0.6 for package docs, (keep existing compat)
