+++ docs/Project.toml
@@ -20,6 +20,7 @@ Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 
 [compat]
+Comrade = "0.6"
 ComradeAHMC = "0.1,0.2"
 ComradeAdaptMCMC = "0.2"
 ComradeBase = "0.3"
