### 2022-08-06 032a415
CompatHelper: bump compat for MeasureBase to 0.13, (keep existing compat)
