### 2022-08-06 d497f2f
update black_hole_image
