### 2022-08-06 #4457 @cescalara
Hi @arfon, I've also completed my review and can recommend Comrade.jl for publication!
