### 2022-08-09 da1a1b3
plot bug fix and add more tests
