+++ test/Core/gains.jl
@@ -1,5 +1,6 @@
 using Distributions
 using Plots
+using Tables
 
 @testset "gains" begin
     _,vis, amp, lcamp, cphase = load_data()
@@ -39,6 +40,34 @@ using Plots
     c1 = caltable(gm)
     c2 = caltable(gcache, ga)
     c3 = caltable(amp, ga)
+
+    @testset "caltable test" begin
+        @test Tables.istable(typeof(c1))
+        @test Tables.rowaccess(typeof(c1))
+        @test Tables.rows(c1) === c1
+        @test Tables.columnaccess(c1)
+        clmns = Tables.columns(c1)
+        @test clmns[1] == Comrade.scantimes(c1)
+        @test Bool(prod(skipmissing(Tables.matrix(clmns)[:,begin+1:end]) .== skipmissing(Comrade.gmat(c1))))
+        @test c1.time == Comrade.scantimes(c1)
+        @test c1.time == Tables.getcolumn(c1, 1)
+        @test c1.AA == Tables.getcolumn(c1, :AA)
+        @test c1.AA == Tables.getcolumn(c1, 2)
+        @test Tables.columnnames(c1) == [:time, sort(stations(amp))...]
+
+        c1row = first(c1)
+        @test eltype(c1) == typeof(c1row)
+        @test c1row.time == c1.time[1]
+        @test c1row.AA == c1.AA[1]
+        @test Tables.getcolumn(c1row, :AA) == c1.AA[1]
+        @test Tables.getcolumn(c1row, :time) == c1.time[1]
+        @test Tables.getcolumn(c1row, 2) == c1.AA[1]
+        @test Tables.getcolumn(c1row, 1) == c1.time[1]
+        @test propertynames(c1) == propertynames(c1row) == [:time, sort(stations(amp))...]
+
+
+    end
+
 
     @test prod(skipmissing(Comrade.gmat(c3) .== Comrade.gmat(c1)))
     @test prod(skipmissing(Comrade.gmat(c3) .== Comrade.gmat(c2)))
