+++ Project.toml
@@ -72,7 +72,7 @@ ForwardDiff = "0.9, 0.10"
 HypercubeTransform = "0.3"
 KeywordCalls = "0.2"
 MeasureBase = "0.7, 0.8, 0.9"
-MeasureTheory = "0.15, 0.16"
+MeasureTheory = "0.15, 0.16, 0.17"
 NFFT = "0.10, 0.11, 0.12"
 NamedTupleTools = "0.13, 0.14"
 PaddedViews = "0.5"
