### 2022-08-12 #4457 @arfon
> Hi @arfon, I've also completed my review and can recommend Comrade.jl for publication!

Awesome news, thanks so much @cescalara!
