+++ examples/imaging.jl
@@ -76,9 +76,9 @@ prior = (
 
 
 mms = GModel(damp, fovx, fovy, nx, ny)
-lklhd = RadioLikelihood(damp, dcphase)
+lklhd = RadioLikelihood(mms, damp, dcphase)
 
-post = Posterior(lklhd, prior, mms)
+post = Posterior(lklhd, prior)
 
 tpost = asflat(post)
 
@@ -89,7 +89,7 @@ using Zygote
 f = OptimizationFunction(tpost, Optimization.AutoZygote())
 prob = OptimizationProblem(f, rand(ndim) .- 0.5, nothing)
 ℓ = logdensityof(tpost)
-sol = solve(prob, LBFGS(); maxiters=6000, callback=(x,p)->(@info ℓ(x); false), g_tol=1e-1)
+sol = solve(prob, LBFGS(); maxiters=2000, callback=(x,p)->(@info ℓ(x); false), g_tol=1e-1)
 
 xopt = transform(tpost, sol)
 
