+++ src/bayes/bayes.jl
@@ -10,7 +10,6 @@ using ValueShapes: NamedTupleDist
 struct Posterior{L,P}
     lklhd::L
     prior::P
-    model::F
 end
 
 """
