+++ src/models/combinators.jl
@@ -163,20 +163,20 @@ end
 #     _combinatorvis(visanalytic(M1), visanalytic(M2), uv_combinator(model), model, u, v, t, ν, cache)
 # end
 
-@inline function _visibilities(model::M, u::AbstractArray, v::AbstractArray, args...) where {M <: CompositeModel}
-    return _visibilities(visanalytic(M), model, u, v, args...)
-end
+# @inline function _visibilities(model::M, u::AbstractArray, v::AbstractArray, args...) where {M <: CompositeModel}
+#     return _visibilities(visanalytic(M), model, u, v, args...)
+# end
 
 
-@inline function _visibilities(::NotAnalytic, model::CompositeModel, u::AbstractArray, v::AbstractArray, args...)
-    f = uv_combinator(model)
-    return f.(_visibilities(model.m1, u, v), _visibilities(model.m2, u, v))
+@inline function _visibilities(model::AddModel, u::AbstractArray, v::AbstractArray, args...)
+    #f = uv_combinator(model)
+    return _visibilities(model.m1, u, v) .+ _visibilities(model.m2, u, v)
 end
 
-@inline function _visibilities(::IsAnalytic, model::CompositeModel, u::AbstractArray, v::AbstractArray, args...)
-    f = uv_combinator(model)
-    return f.(visibility_point.(Ref(model.m1), u, v), visibility_point.(Ref(model.m2), u, v))
-end
+# @inline function _visibilities(::IsAnalytic, model::CompositeModel, u::AbstractArray, v::AbstractArray, args...)
+#     f = uv_combinator(model)
+#     return f.(visibility_point.(Ref(model.m1), u, v), visibility_point.(Ref(model.m2), u, v))
+# end
 
 
 
@@ -284,6 +284,14 @@ function intensitymap!(::NotAnalytic, sim::IntensityMap, model::ConvolvedModel, 
         sim[I] = real(vis[I])/(nx*ny)
     end
 end
+
+#ChainRulesCore.@non_differentiable getproperty(m::ConvolvedModel, x::Symbol)
+
+@inline function _visibilities(model::ConvolvedModel, u::AbstractArray, v::AbstractArray, args...)
+    #f = uv_combinator(model)
+    return _visibilities(model.m1, u, v).*_visibilities(model.m2, u, v)
+end
+
 
 
 # function _combinatorvis(::IsAnalytic, ::IsAnalytic, f::F, m, u, v, t, ν, cache) where {F}
