+++ README.md
@@ -4,6 +4,8 @@
 [![Build Status](https://github.com/ptiede/Comrade.jl/workflows/CI/badge.svg)](https://github.com/ptiede/Comrade.jl/actions)
 [![Coverage](https://codecov.io/gh/ptiede/Comrade.jl/branch/main/graph/badge.svg)](https://codecov.io/gh/ptiede/Comrade.jl)
 [![ColPrac: Contributor's Guide on Collaborative Practices for Community Packages](https://img.shields.io/badge/ColPrac-Contributor's%20Guide-blueviolet)](https://github.com/SciML/ColPrac)
+[![DOI](https://zenodo.org/badge/320609516.svg)](https://zenodo.org/badge/latestdoi/320609516)
+
 
 Composable Modeling of Radio Emission
 
