### 2022-08-12 #4457 @ptiede
First, @cescalara and @mileslucas thanks for all your hard work the package is much better now!

The most recent version of Comrade has all the requested updates during the referee process. I have gone ahead and created a zenodo release for the most recent version the doi is:

https://doi.org/10.5281/zenodo.6977819

Let me know if you need anything else!
