### 2022-08-13 #4457 @arfon
@ptiede - it looks like `EHTCIV`, `EHTCVI`, and `themis` are all listed twice in your bibtex file. Could you please delete the duplicates? This messes up our production pipeline (as you can see above).
