### 2022-08-13 bc042fa
update imaging example
