+++ paper/paper.bib
@@ -213,65 +213,6 @@ archivePrefix = {arXiv},
 }
 
 
-
-@ARTICLE{EHTCIV,
-       author = {{Event Horizon Telescope Collaboration}},
-        title = "{First M87 Event Horizon Telescope Results. IV. Imaging the Central Supermassive Black Hole}",
-      journal = {\apjl},
-     keywords = {black hole physics, galaxies: individual: M87, galaxies: jets, techniques: high angular resolution, techniques: image processing, techniques: interferometric, Astrophysics - Astrophysics of Galaxies, Astrophysics - Instrumentation and Methods for Astrophysics, General Relativity and Quantum Cosmology},
-         year = 2019,
-        month = apr,
-       volume = {875},
-       number = {1},
-          eid = {L4},
-        pages = {L4},
-          doi = {10.3847/2041-8213/ab0e85},
-archivePrefix = {arXiv},
-       eprint = {1906.11241},
- primaryClass = {astro-ph.GA},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2019ApJ...875L...4E},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-
-
-
-
-@ARTICLE{EHTCVI,
-       author = {{Event Horizon Telescope Collaboration}},
-        title = "{First M87 Event Horizon Telescope Results. VI. The Shadow and Mass of the Central Black Hole}",
-      journal = {\apjl},
-     keywords = {black hole physics, galaxies: individual: M87, gravitation, techniques: high angular resolution, techniques: interferometric, Astrophysics - Astrophysics of Galaxies, Astrophysics - High Energy Astrophysical Phenomena, General Relativity and Quantum Cosmology},
-         year = 2019,
-        month = apr,
-       volume = {875},
-       number = {1},
-          eid = {L6},
-        pages = {L6},
-          doi = {10.3847/2041-8213/ab1141},
-archivePrefix = {arXiv},
-       eprint = {1906.11243},
- primaryClass = {astro-ph.GA},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2019ApJ...875L...6E},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@ARTICLE{themis,
-       author = {{Broderick}, Avery E. and {Gold}, Roman and {Karami}, Mansour and {Preciado-L{\'o}pez}, Jorge A. and {Tiede}, Paul and {Pu}, Hung-Yi and et al.},
-        title = "{THEMIS: A Parameter Estimation Framework for the Event Horizon Telescope}",
-      journal = {\apj},
-     keywords = {565, 1647, 1858, 1769, 98},
-         year = 2020,
-        month = jul,
-       volume = {897},
-       number = {2},
-          eid = {139},
-        pages = {139},
-          doi = {10.3847/1538-4357/ab91a4},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2020ApJ...897..139B},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
 @misc{gal,
   author = {Saurabh},
   title = {Galifray: Geometric Modelling and Parameter Estimation using Black hole Images for EHT including VLBI},
