+++ paper/paper.bib
@@ -213,28 +213,9 @@ archivePrefix = {arXiv},
 }
 
 
-@misc{gal,
-  author = {Saurabh},
-  title = {Galifray: Geometric Modelling and Parameter Estimation using Black hole Images for EHT including VLBI},
-  year = {2022},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/Relativist1/Gallifray}
-}
-
-@misc{in,
-  author = {{Plavin}, Alexander},
-  title = {InterferometricModels},
-  year = {2022},
-  publisher = {GitLab},
-  journal = {GitLab repository},
-  url = {https://gitlab.com/aplavin/InterferometricModels.jl}
-}
-
-
 
 @ARTICLE{EHTCIV,
-       author = {{Event Horizon Telescope Collaboration} and {Akiyama}, Kazunori and {Alberdi}, Antxon and {Alef}, Walter and {Asada}, Keiichi and {Azulay}, Rebecca and {Baczko}, Anne-Kathrin and {Ball}, David and {Balokovi{\'c}}, Mislav and {Barrett}, John and {Bintley}, Dan and {Blackburn}, Lindy and {Boland}, Wilfred and {Bouman}, Katherine L. and {Bower}, Geoffrey C. and {Bremer}, Michael and {Brinkerink}, Christiaan D. and {Brissenden}, Roger and {Britzen}, Silke and {Broderick}, Avery E. and {Broguiere}, Dominique and {Bronzwaer}, Thomas and {Byun}, Do-Young and {Carlstrom}, John E. and {Chael}, Andrew and {Chan}, Chi-kwan and {Chatterjee}, Shami and {Chatterjee}, Koushik and {Chen}, Ming-Tang and {Chen}, Yongjun and {Cho}, Ilje and {Christian}, Pierre and {Conway}, John E. and {Cordes}, James M. and {Crew}, Geoffrey B. and {Cui}, Yuzhu and {Davelaar}, Jordy and {De Laurentis}, Mariafelicia and {Deane}, Roger and {Dempsey}, Jessica and {Desvignes}, Gregory and {Dexter}, Jason and {Doeleman}, Sheperd S. and {Eatough}, Ralph P. and {Falcke}, Heino and {Fish}, Vincent L. and {Fomalont}, Ed and {Fraga-Encinas}, Raquel and {Freeman}, William T. and {Friberg}, Per and {Fromm}, Christian M. and {G{\'o}mez}, Jos{\'e} L. and {Galison}, Peter and {Gammie}, Charles F. and {Garc{\'\i}a}, Roberto and {Gentaz}, Olivier and {Georgiev}, Boris and {Goddi}, Ciriaco and {Gold}, Roman and {Gu}, Minfeng and {Gurwell}, Mark and {Hada}, Kazuhiro and {Hecht}, Michael H. and {Hesper}, Ronald and {Ho}, Luis C. and {Ho}, Paul and {Honma}, Mareki and {Huang}, Chih-Wei L. and {Huang}, Lei and {Hughes}, David H. and {Ikeda}, Shiro and {Inoue}, Makoto and {Issaoun}, Sara and {James}, David J. and {Jannuzi}, Buell T. and {Janssen}, Michael and {Jeter}, Britton and {Jiang}, Wu and {Johnson}, Michael D. and {Jorstad}, Svetlana and {Jung}, Taehyun and {Karami}, Mansour and {Karuppusamy}, Ramesh and {Kawashima}, Tomohisa and {Keating}, Garrett K. and {Kettenis}, Mark and {K
[…]
