+++ paper/paper.bib
@@ -256,21 +256,6 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{themis,
-       author = {{Broderick}, Avery E. and {Gold}, Roman and {Karami}, Mansour and {Preciado-L{\'o}pez}, Jorge A. and {Tiede}, Paul and {Pu}, Hung-Yi and et al.},
-        title = "{THEMIS: A Parameter Estimation Framework for the Event Horizon Telescope}",
-      journal = {\apj},
-     keywords = {565, 1647, 1858, 1769, 98},
-         year = 2020,
-        month = jul,
-       volume = {897},
-       number = {2},
-          eid = {139},
-        pages = {139},
-          doi = {10.3847/1538-4357/ab91a4},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2020ApJ...897..139B},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
 
 @misc{gal,
   author = {Saurabh},
@@ -290,6 +275,22 @@ archivePrefix = {arXiv},
   url = {https://gitlab.com/aplavin/InterferometricModels.jl}
 }
 
+
+@ARTICLE{themis,
+       author = {{Broderick}, Avery E. and {Gold}, Roman and {Karami}, Mansour and {Preciado-L{\'o}pez}, Jorge A. and {Tiede}, Paul and {Pu}, Hung-Yi and {Akiyama}, Kazunori and {Alberdi}, Antxon and {Alef}, Walter and {Asada}, Keiichi and {Azulay}, Rebecca and {Baczko}, Anne-Kathrin and {Balokovi{\'c}}, Mislav and {Barrett}, John and {Bintley}, Dan and {Blackburn}, Lindy and {Boland}, Wilfred and {Bouman}, Katherine L. and {Bower}, Geoffrey C. and {Bremer}, Michael and {Brinkerink}, Christiaan D. and {Brissenden}, Roger and {Britzen}, Silke and {Broguiere}, Dominique and {Bronzwaer}, Thomas and {Byun}, Do-Young and {Carlstrom}, John E. and {Chael}, Andrew and {Chatterjee}, Shami and {Chatterjee}, Koushik and {Chen}, Ming-Tang and {Chen}, Yongjun and {Cho}, Ilje and {Conway}, John E. and {Cordes}, James M. and {Crew}, Geoffrey B. and {Cui}, Yuzhu and {Davelaar}, Jordy and {De Laurentis}, Mariafelicia and {Deane}, Roger and {Dempsey}, Jessica and {Desvignes}, Gregory and {Doeleman}, Sheperd S. and {Eatough}, Ralph P. and {Falcke}, Heino and {Fish}, Vincent L. and {Fomalont}, Ed and {Fraga-Encinas}, Raquel and {Friberg}, Per and {Fromm}, Christian M. and {Galison}, Peter and {Gammie}, Charles F. and {Garc{\'\i}a}, Roberto and {Gentaz}, Olivier and {Georgiev}, Boris and {Goddi}, Ciriaco and {G{\'o}mez}, Jos{\'e} L. and {Gu}, Minfeng and {Gurwell}, Mark and {Hada}, Kazuhiro and {Hecht}, Michael H. and {Hesper}, Ronald and {Ho}, Luis C. and {Ho}, Paul and {Honma}, Mareki and {Huang}, Chih-Wei L. and {Huang}, Lei and {Hughes}, David H. and {Inoue}, Makoto and {Issaoun}, Sara and {James}, 
[…]
