### 2022-08-14 82a30ea
fix xlims in caltable plot
