### 2022-08-14 b2bab9b
add centroid testing
