+++ Project.toml
@@ -1,7 +1,7 @@
 name = "Comrade"
 uuid = "99d987ce-9a1e-4df8-bc0b-1ea019aa547b"
 authors = ["Paul Tiede <ptiede91@gmail.com>"]
-version = "0.6.4"
+version = "0.6.5"
 
 [deps]
 AbstractDifferentiation = "c29ec348-61ec-40c8-8164-b8c60e9d9f3d"
@@ -59,7 +59,7 @@ ArgCheck = "2"
 AstroTime = "0.6,0.7"
 BasicInterpolators = "0.6"
 ChainRulesCore = "1"
-ComradeBase = "^0.3.1"
+ComradeBase = "^0.3.4"
 DensityInterface = "0.4"
 Distributions = "0.24,0.25"
 DocStringExtensions = "0.6,0.7,0.8, 0.9"
