### 2022-08-14 #4457 @arfon
@ptiede – found a couple of typos in the paper: https://github.com/ptiede/Comrade.jl/pull/170
